Crate gix

This crate provides the Repository abstraction which serves as a hub into all the functionality of git.

It’s powerful and won’t sacrifice performance while still increasing convenience compared to using the sub-crates individually. Sometimes it may hide complexity under the assumption that the performance difference doesn’t matter for all but the fewest tools out there, which would be using the underlying crates directly or file an issue.

The Trust Model

It is very simple - based on the ownership of the repository compared to the user of the current process Trust is assigned. This can be overridden as well. Further, git configuration files track their trust level per section based on and sensitive values like paths to executables or certain values will be skipped if they are from a source that isn’t fully trusted.

That way, data can safely be obtained without risking to execute untrusted executables.

Note that it’s possible to let gix act like git or git2 by setting the open::Options::bail_if_untrusted() option.

The prelude and extensions

With use git_repository::prelude::* you should be ready to go as it pulls in various extension traits to make functionality available on objects that may use it.

The method signatures are still complex and may require various arguments for configuration and cache control.

Most extensions to existing objects provide an obj_with_extension.attach(&repo).an_easier_version_of_a_method() for simpler call signatures.

ThreadSafe Mode

By default, the Repository isn’t Sync and thus can’t be used in certain contexts which require the Sync trait.

To help with this, convert it with .into_sync() into a ThreadSafeRepository.

Object-Access Performance

Accessing objects quickly is the bread-and-butter of working with git, right after accessing references. Hence it’s vital to understand which cache levels exist and how to leverage them.

When accessing an object, the first cache that’s queried is a memory-capped LRU object cache, mapping their id to data and kind. It has to be specifically enabled a Repository. On miss, the object is looked up and if a pack is hit, there is a small fixed-size cache for delta-base objects.

In scenarios where the same objects are accessed multiple times, the object cache can be useful and is to be configured specifically using the object_cache_size(…) method.

Use the cache-efficiency-debug cargo feature to learn how efficient the cache actually is - it’s easy to end up with lowered performance if the cache is not hit in 50% of the time.

Terminology

WorkingTree and WorkTree

When reading the documentation of the canonical gix-worktree program one gets the impression work tree and working tree are used interchangeably. We use the term work tree only and try to do so consistently as its shorter and assumed to be the same.

Plumbing Crates

To make using sub-crates and their types easier, these are re-exported into the root of this crate. Here we list how to access nested plumbing crates which are otherwise harder to discover:

git_repository::

[odb]

[pack][odb::pack]

[protocol]

[transport][protocol::transport]

[packetline][protocol::transport::packetline]

libgit2 API to gix

This doc-aliases are used to help finding methods under a possibly changed name. Just search in the docs. Entering git2 into the search field will also surface all methods with such annotations.

What follows is a list of methods you might be missing, along with workarounds if available.

git2::Repository::open_bare() ➡ ❌ - use open() and discard if it is not bare.

git2::build::CheckoutBuilder::disable_filters() ➡ ❌ (filters are always applied during checkouts)

git2::Repository::submodule_status() ➡ Submodule::state() - status provides more information and conveniences though, and an actual worktree status isn’t performed.

Integrity checks

git2 by default performs integrity checks via strict_hash_verification() and strict_object_creation which gitoxide currently does not have.

Feature Flags

Re-exports

pub use gix_actor as actor;

pub use gix_attributes as attrs;

pub use gix_command as command;

pub use gix_commitgraph as commitgraph;

pub use gix_credentials as credentials;

pub use gix_date as date;

pub use gix_dir as dir;

pub use gix_features as features;

pub use gix_fs as fs;

pub use gix_glob as glob;

pub use gix_hash as hash;

pub use gix_hashtable as hashtable;

pub use gix_ignore as ignore;

pub use gix_lock as lock;

pub use gix_negotiate as negotiate;

pub use gix_object as objs;

pub use gix_object::bstr;

pub use gix_odb as odb;

pub use gix_prompt as prompt;

pub use gix_protocol as protocol;

pub use gix_ref as refs;

pub use gix_refspec as refspec;

pub use gix_revwalk as revwalk;

pub use gix_sec as sec;

pub use gix_tempfile as tempfile;

pub use gix_trace as trace;

pub use gix_traverse as traverse;

pub use gix_url as url;

pub use gix_utils as utils;

pub use gix_validate as validate;

Modules

clone

commit

config

create

diff

dirwalk

discover

env

Utilities to handle program arguments and other values of interest.

filter

lower-level access to filters which are applied to create working tree checkouts or to ‘clean’ working tree contents for storage in git.

head

id

index

Feature Flags

init

interrupt

Process-global interrupt handling

object

open

parallel

Run computations in parallel, or not based the parallel feature toggle.

path

pathspec

Pathspec plumbing and abstractions

prelude

progress

push

reference

remote

repository

revision

Revisions is the generalized notion of a commit.

shallow

state

Not to be confused with ‘status’.

status

submodule

Submodule plumbing and abstractions

tag

threading

Type definitions for putting shared ownership and synchronized mutation behind the threading feature toggle.

worktree

Structs

AttributeStack

A utility to access .gitattributes and .gitignore information efficiently.

Blob

A blob along with access to its owning repository.

Commit

A decoded commit object with access to its owning repository.

Head

The head reference, as created from looking at .git/HEAD, able to represent all of its possible states.

Id

An ObjectId with access to a repository.

Object

A decoded object with a reference to its owning repository.

ObjectDetached

A detached, self-contained object, without access to its source repository.

Pathspec

A utility to make matching against pathspecs simple.

PathspecDetached

Like Pathspec, but without a Repository reference and with minimal API.

Reference

A reference that points to an object or reference, with access to its source repository.

Remote

A remote which represents a way to interact with hosts for remote clones of the parent repository.

Repository

A thread-local handle to interact with a repository from a single thread.

Submodule

A stand-in for the submodule of a particular name.

Tag

A decoded tag object with access to its owning repository.

ThreadSafeRepository

An instance with access to everything a git repository entails, best imagined as container implementing Sync + Send for most for system resources required to interact with a git repository which are loaded in once the instance is created.

Tree

A decoded tree object with access to its owning repository.

Url

A URL with support for specialized git related capabilities.

Worktree

A worktree checkout containing the files of the repository in consumable form.

oid

A borrowed reference to a hash identifying objects.

Enums

ObjectId

An owned hash identifying objects, most commonly Sha1

Traits

Count

A thread-safe read-only counter, with unknown limits.

DynNestedProgress

An object-safe trait for describing hierarchical progress.

NestedProgress

A trait for describing hierarchical progress.

Progress

A trait for describing non-hierarchical progress.

Functions

discover

See ThreadSafeRepository::discover(), but returns a Repository instead.

init

See ThreadSafeRepository::init(), but returns a Repository instead.

init_bare

See ThreadSafeRepository::init(), but returns a Repository instead.

open

See ThreadSafeRepository::open(), but returns a Repository instead.

open_opts

See ThreadSafeRepository::open_opts(), but returns a Repository instead.

prepare_clone

Create a platform for configuring a clone with main working tree from url to the local path, using default options for opening it (but amended with using configuration from the git installation to ensure all authentication options are honored).

prepare_clone_bare

Create a platform for configuring a bare clone from url to the local path, using default options for opening it (but amended with using configuration from the git installation to ensure all authentication options are honored).

Type Aliases

OdbHandle

A handle for finding objects in an object database, abstracting away caches for thread-local use.

OdbHandleArc

A handle for finding objects in an object database, abstracting away caches for moving across threads.

RefStore

The standard type for a store to handle git references.