gix::clone

Struct PrepareFetch

pub struct PrepareFetch { /* private fields */ }

A utility to collect configuration on how to fetch from a remote and initiate a fetch operation. It will delete the newly created repository on when dropped without successfully finishing a fetch.

Implementations

impl PrepareFetch

Builder

pub fn configure_connection( self, f: impl FnMut(&mut Connection<'_, '_, Box<dyn Transport + Send>>) -> Result<(), Box<dyn Error + Send + Sync>> + 'static, ) -> Self

Set a callback to use for configuring the connection to use right before connecting to the remote.

It is most commonly used for custom configuration.

pub fn with_fetch_options(self, opts: Options) -> Self

Set additional options to adjust parts of the fetch operation that are not affected by the git configuration.

impl PrepareFetch

Builder

pub fn configure_remote( self, f: impl FnMut(Remote<'_>) -> Result<Remote<'_>, Box<dyn Error + Send + Sync>> + 'static, ) -> Self

Use f to apply arbitrary changes to the remote that is about to be used to fetch a pack.

The passed in remote will be un-named and pre-configured to be a default remote as we know it from git-clone. It is not yet present in the configuration of the repository, but each change it will eventually be written to the configuration prior to performing a the fetch operation, all changes done in f() will be persisted.

It can also be used to configure additional options, like those for fetching tags. Note that with_fetch_tags() should be called here to configure the clone as desired. Otherwise, a clone is configured to be complete and fetches all tags, not only those reachable from all branches.

pub fn with_remote_name(self, name: impl Into<BString>) -> Result<Self, Error>

Set the remote’s name to the given value after it was configured using the function provided via configure_remote().

If not set here, it defaults to origin or the value of clone.defaultRemoteName.

pub fn with_shallow(self, shallow: Shallow) -> Self

Make this clone a shallow one with the respective choice of shallow-ness.

pub fn with_in_memory_config_overrides( self, values: impl IntoIterator<Item = impl Into<BString>>, ) -> Self

Apply the given configuration values right before readying the actual fetch from the remote. The configuration is marked with source API, and will not be written back, it’s retained only in memory.

pub fn with_ref_name<'a, Name, E>(self, name: Option<Name>) -> Result<Self, E>
where Name: TryInto<&'a PartialNameRef, Error = E>,

Set the name of the reference to check out, instead of the remote HEAD. If None, the HEAD will be used, which is the default.

Note that name should be a partial name like main or feat/one, but can be a full ref name. If a branch on the remote matches, it will automatically be retrieved even without a refspec.

impl PrepareFetch

Consumption

pub fn persist(self) -> Repository

Persist the contained repository as is even if an error may have occurred when fetching from the remote.

impl PrepareFetch

Instantiation

pub fn new<Url, E>( url: Url, path: impl AsRef<Path>, kind: Kind, create_opts: Options, open_opts: Options, ) -> Result<Self, Error>
where Url: TryInto<Url, Error = E>, Error: From<E>,

Create a new repository at path with create_opts which is ready to clone from url, possibly after making additional adjustments to configuration and settings.

Note that this is merely a handle to perform the actual connection to the remote, and if any of it fails the freshly initialized repository will be removed automatically as soon as this instance drops.

Deviation

Similar to git, a missing user name and email configuration is not terminal and we will fill it in with dummy values. However, instead of deriving values from the system, ours are hardcoded to indicate what happened.