gix

Module config

Modules

abbrev

attribute_stack

boolean

checkout

checkout_options

color

command_context

credential_helpers

diff

duration

encoding

exclude_stack

file

A high level wrapper around a single or multiple git-config file, for reading and mutation.

integer

key

lock_timeout

lookup

merge

overrides

parse

This module handles parsing a git-config file. Generally speaking, you want to use a higher abstraction such as File unless you have some explicit reason to work with events instead.

path

protocol

refs_namespace

refspec

remote

section

set_value

source

ssh_connect_options

ssl_version

stat_options

string

time

transport

tree

The tree of supported configuration values for use in config_overrides or for validating and transforming well-known configuration values.

unsigned_integer

url

value

Structs

Boolean

Any value that can be interpreted as a boolean.

Color

Any value that may contain a foreground color, background color, a collection of color (text) modifiers, or a combination of any of the aforementioned values, like red or brightgreen.

CommitAutoRollback

A utility structure created by SnapshotMut::commit_auto_rollback() that restores the previous configuration on drop.

File

High level git-config reader and writer.

Integer

Any value that can be interpreted as an integer.

KeyRef

An unvalidated parse result of parsing input like remote.origin.url or core.bare.

Path

Any value that can be interpreted as a path to a resource on disk.

Snapshot

A platform to access configuration values as read from disk.

SnapshotMut

A platform to access configuration values and modify them in memory, while making them available when this platform is dropped as form of auto-commit. Note that the values will only affect this instance of the parent repository, and not other clones that may exist.

Tree

The root of the configuration tree, suitable to discover all sub-sections at runtime or compile time.

Enums

Error

The error returned when failing to initialize the repository configuration.

Source

A list of known sources for git configuration in order of ascending precedence.

Traits

AsKey

Parse parts of a Git configuration key, like remote.origin.url or core.bare.

Functions

credential_helpers

Returns the configuration for all git-credential helpers from trusted configuration that apply to the given url along with an action preconfigured to invoke the cascade with to retrieve it. This includes url which may be altered to contain a user-name as configured.