gix::config::duration

Type Alias Error

pub type Error = Error<Error, 'd', 'i'>;

The error produced when failing to parse durations (in milliseconds).

Aliased Type

struct Error {
    pub key: BString,
    pub value: Option<BString>,
    pub environment_override: Option<&'static str>,
    pub source: Option<Error>,
}

Fields
key: BString

The configuration key that contained the value.

value: Option<BString>

The value that was assigned to key.

environment_override: Option<&'static str>

The associated environment variable that would override this value.

source: Option<Error>

The source of the error if there was one.