gix::config

Enum Source

pub enum Source {
    GitInstallation,
    System,
    Git,
    User,
    Local,
    Worktree,
    Env,
    Cli,
    Api,
    EnvOverride,
}

A list of known sources for git configuration in order of ascending precedence.

This means values from the first one will be overridden by values in the second one, and so forth. Note that included files via include.path and includeIf.<condition>.path inherit their source.

Variants

GitInstallation

A special configuration file that ships with the git installation, and is thus tied to the used git binary.

System

System-wide configuration path. This is defined as $(prefix)/etc/gitconfig (where prefix is the git-installation directory).

Git

A platform defined location for where a userโs git application configuration should be located. If $XDG_CONFIG_HOME is not set or empty, $HOME/.config/git/config will be used on unix.

User

This is usually ~/.gitconfig on unix.

Local

The configuration of the repository itself, located in .git/config.

Worktree

Configuration specific to a worktree as created with git worktree and typically located in $GIT_DIR/config.worktree if extensions.worktreeConfig is enabled.

Env

Values parsed from the environment using GIT_CONFIG_COUNT, GIT_CONFIG_KEY_N and GIT_CONFIG_VALUE_N where N is incremented from 0 up to the value of GIT_CONFIG_COUNT.

Cli

Values set from the command-line, typically controlled by the user running a program.

Api

Entirely internal from a programmatic source, and can be used to have (near final) say in configuration values.

EnvOverride

Values obtained from specific environment variables that override values in the git configuration.

For example, HTTP_PROXY overrides http.proxy, no matter where it is specified, and thus controls the value similar to how itโs done in git.

Implementations

impl Source

pub const fn kind(self) -> Kind

Return true if the source indicates a location within a file of a repository.

pub fn storage_location( self, env_var: &mut dyn FnMut(&str) -> Option<OsString>, ) -> Option<Cow<'static, Path>>

Returns the location at which a file of this type would be stored, or None if there is no notion of persistent storage for this source, with env_var to obtain environment variables. Note that the location can be relative for repository-local sources like Local and Worktree, and the caller has to known which base it is relative to, namely the common_dir in the Local case and the git_dir in the Worktree case. Be aware that depending on environment overrides, multiple scopes might return the same path, which should only be loaded once nonetheless.

With env_var it becomes possible to prevent accessing environment variables entirely to comply with gix-sec permissions for example.