gix::config

Module file

A high level wrapper around a single or multiple git-config file, for reading and mutation.

Modules

includes

init

rename_section

section

set_raw_value

Structs

Metadata

Additional information about a section.

MultiValueMut

An intermediate representation of a mutable multivar obtained from a File.

Section

A section in a git-config file, like [core] or [remote "origin"], along with all of its keys.

SectionId

The section ID is a monotonically increasing ID used to refer to section bodies. This value does not imply any ordering between sections, as new sections with higher section IDs may be in between lower ID sections after File mutation.

SectionMut

A opaque type that represents a mutable reference to a section.

ValueMut

An intermediate representation of a mutable value obtained from a File.