gix::config::file

Struct SectionMut

pub struct SectionMut<'a, 'event> { /* private fields */ }

A opaque type that represents a mutable reference to a section.

Implementations

impl<'event> SectionMut<'_, 'event>

Mutating methods.

pub fn push<'b>( &mut self, value_name: ValueName<'event>, value: Option<&'b BStr>, ) -> &mut SectionMut<'_, 'event>

Adds an entry to the end of this section name value_name and value. If value is None, no equal sign will be written leaving just the key. This is useful for boolean values which are true if merely the key exists.

pub fn push_with_comment<'b, 'c>( &mut self, value_name: ValueName<'event>, value: Option<&'b BStr>, comment: impl Into<&'c BStr>, ) -> &mut SectionMut<'_, 'event>

Adds an entry to the end of this section name value_name and value. If value is None, no equal sign will be written leaving just the key. This is useful for boolean values which are true if merely the key exists. comment has to be the text to put right after the value and behind a # character. Note that newlines are silently transformed into spaces.

pub fn pop(&mut self) -> Option<(ValueName<'_>, Cow<'event, BStr>)>

Removes all events until a key value pair is removed. This will also remove the whitespace preceding the key value pair, if any is found.

pub fn set( &mut self, value_name: ValueName<'event>, value: &BStr, ) -> Option<Cow<'event, BStr>>

Sets the last key value pair if it exists, or adds the new value. Returns the previous value if it replaced a value, or None if it adds the value.

pub fn set_trust(&mut self, trust: Trust) -> &mut SectionMut<'_, 'event>

Set the trust level in the meta-data of this section to trust.

pub fn remove(&mut self, value_name: &str) -> Option<Cow<'event, BStr>>

Removes the latest value by key and returns it, if it exists.

pub fn push_newline(&mut self) -> &mut SectionMut<'_, 'event>

Adds a new line event. Note that you don’t need to call this unless you’ve disabled implicit newlines.

pub fn newline(&self) -> &BStr

Return the newline used when calling push_newline().

pub fn set_implicit_newline(&mut self, on: bool) -> &mut SectionMut<'_, 'event>

Enables or disables automatically adding newline events after adding a value. This is enabled by default.

pub fn set_leading_whitespace( &mut self, whitespace: Option<Cow<'event, BStr>>, ) -> &mut SectionMut<'_, 'event>

Sets the exact whitespace to use before each newly created key-value pair, with only whitespace characters being permissible.

The default is 2 tabs. Set to None to disable adding whitespace before a key value.

Panics

If non-whitespace characters are used. This makes the method only suitable for validated or known input.

pub fn leading_whitespace(&self) -> Option<&BStr>

Returns the whitespace this section will insert before the beginning of a key, if any.

pub fn separator_whitespace(&self) -> (Option<&BStr>, Option<&BStr>)

Returns the whitespace to be used before and after the = between the key and the value.

For example, k = v will have (Some(" "), Some(" ")), whereas k=\tv will have (None, Some("\t")).