gix::config

Function credential_helpers

pub fn credential_helpers(
    url: Url,
    config: &File<'_>,
    is_lenient_config: bool,
    filter: impl FnMut(&Metadata) -> bool,
    environment: Environment,
    use_http_path: bool,
) -> Result<(Cascade, Action, Options<'static>), Error>

Returns the configuration for all git-credential helpers from trusted configuration that apply to the given url along with an action preconfigured to invoke the cascade with to retrieve it. This includes url which may be altered to contain a user-name as configured.

These can be invoked to obtain credentials. Note that the url is expected to be the one used to connect to a remote, and thus should already have passed the url-rewrite engine.

config

the configuration to obtain credential helper configuration from.

is_lenient_config

if true, minor configuration errors will be ignored.

filter

A way to choose which sections in config can be trusted. This is important as we will execute programs from the paths contained within.

environment

Determines how environment variables can be used.

Actually used are GIT_* and SSH_* environment variables to configure git prompting capabilities.

use_http_path

Typically, this should be false to let the url configuration decide if the value should be enabled.

If false, credentials are effectively per host.

Deviation

Invalid urls can’t be used to obtain credential helpers as they are rejected early when creating a valid url here.

Parsed urls will automatically drop the port if it’s the default, i.e. http://host:80 becomes http://host when parsed. This affects the prompt provided to the user, so that git will use the verbatim url, whereas we use http://host.

Upper-case scheme and host will be lower-cased automatically when parsing into a url, so prompts differ compared to git.

A difference in prompt might affect the matching of getting existing stored credentials, and it’s a question of this being a feature or a bug.