gix::config::key

Struct Error

pub struct Error<E: Error + Send + Sync + 'static, const PREFIX: char, const SUFFIX: char> {
    pub key: BString,
    pub value: Option<BString>,
    pub environment_override: Option<&'static str>,
    pub source: Option<E>,
}

A generic error suitable to produce decent messages for all kinds of configuration errors with config-key granularity.

This error is meant to be reusable and help produce uniform error messages related to parsing any configuration key.

Fields
key: BString

The configuration key that contained the value.

value: Option<BString>

The value that was assigned to key.

environment_override: Option<&'static str>

The associated environment variable that would override this value.

source: Option<E>

The source of the error if there was one.

Implementations

impl<E, const PREFIX: char, const SUFFIX: char> Error<E, PREFIX, SUFFIX>
where E: Error + Send + Sync + 'static,

Initialization

pub fn from_value(key: &'static impl Key, value: BString) -> Self

Instantiate an error with all data from key along with the value of the key.

impl<E, const PREFIX: char, const SUFFIX: char> Error<E, PREFIX, SUFFIX>
where E: Error + Send + Sync + 'static,

Builder

pub fn with_source(self, err: E) -> Self

Attach the given err as source.

pub fn with_value(self, value: BString) -> Self

Attach the given value as value we observed when the error was produced.