gix::config::parse

Struct Events

pub struct Events<'a> {
    pub frontmatter: SmallVec<[Event<'a>; 8]>,
    pub sections: Vec<Section<'a>>,
}

A zero-copy git-config file parser.

This is parser exposes low-level syntactic events from a git-config file. Generally speaking, you’ll want to use File as it wraps around the parser to provide a higher-level abstraction to a git-config file, including querying, modifying, and updating values.

This parser guarantees that the events emitted are sufficient to reconstruct a git-config file identical to the source git-config when writing it.

Differences between a .ini parser

While the git-config format closely resembles the .ini file format, there are subtle differences that make them incompatible. For one, the file format is not well defined, and there exists no formal specification to adhere to.

For concrete examples, some notable differences are:

git-config sections permit subsections via either a quoted string ([some-section "subsection"]) or via the deprecated dot notation ([some-section.subsection]). Successful parsing these section names is not well defined in typical .ini parsers. This parser will handle these cases perfectly.

Comment markers are not strictly defined either. This parser will always and only handle a semicolon or octothorpe (also known as a hash or number sign).

Global properties may be allowed in .ini parsers, but is strictly disallowed by this parser.

Only \t, \n, \b \\ are valid escape characters.

Quoted and semi-quoted values will be parsed (but quotes will be included in event outputs). An example of a semi-quoted value is 5"hello world", which should be interpreted as 5hello world after normalization.

Line continuations via a \ character is supported (inside or outside of quotes)

Whitespace handling similarly follows the git-config specification as closely as possible, where excess whitespace after a non-quoted value are trimmed, and line continuations onto a new line with excess spaces are kept.

Only equal signs (optionally padded by spaces) are valid name/value delimiters.

Note that things such as case-sensitivity or duplicate sections are not handled. This parser is a low level syntactic interpreter and higher level wrappers around this parser, which may or may not be zero-copy, should handle semantic values. This also means that string-like values are not interpreted. For example, hello"world" would be read at a high level as helloworld but this parser will return the former instead, with the extra quotes. This is because it is not the responsibility of the parser to interpret these values, and doing so would necessarily require a copy, which this parser avoids.