gix::config

Module section

Functions

is_trusted

A filter that returns true for meta if the meta-data attached to a configuration section can be trusted. This is either the case if its file is fully trusted, or if it’s a section from a system-wide file.