gix::config

Struct Snapshot

pub struct Snapshot<'repo> {
    pub repo: &'repo Repository,
}

A platform to access configuration values as read from disk.

Note that these values won’t update even if the underlying file(s) change.

Fields
repo: &'repo Repository

The owning repository.

Implementations

impl<'repo> Snapshot<'repo>

Access configuration values, frozen in time, using a key which is a . separated string of up to three tokens, namely section_name.[subsection_name.]value_name, like core.bare or remote.origin.url.

Note that single-value methods always return the last value found, which is the one set most recently in the hierarchy of configuration files, aka ‘last one wins’.

pub fn boolean(&self, key: impl AsKey) -> Option<bool>

Return the boolean at key, or None if there is no such value or if the value can’t be interpreted as boolean.

For a non-degenerating version, use try_boolean(…).

Note that this method takes the most recent value at key even if it is from a file with reduced trust.

pub fn try_boolean(&self, key: impl AsKey) -> Option<Result<bool, Error>>

Like boolean(), but it will report an error if the value couldn’t be interpreted as boolean.

pub fn integer(&self, key: impl AsKey) -> Option<i64>

Return the resolved integer at key, or None if there is no such value or if the value can’t be interpreted as integer or exceeded the value range.

For a non-degenerating version, use try_integer(…).

Note that this method takes the most recent value at key even if it is from a file with reduced trust.

pub fn try_integer(&self, key: impl AsKey) -> Option<Result<i64, Error>>

Like integer(), but it will report an error if the value couldn’t be interpreted as boolean.

pub fn string(&self, key: impl AsKey) -> Option<Cow<'repo, BStr>>

Return the string at key, or None if there is no such value.

Note that this method takes the most recent value at key even if it is from a file with reduced trust.

pub fn trusted_path( &self, key: impl AsKey, ) -> Option<Result<Cow<'repo, Path>, Error>>

Return the trusted and fully interpolated path at key, or None if there is no such value or if no value was found in a trusted file. An error occurs if the path could not be interpolated to its final value.

pub fn trusted_program(&self, key: impl AsKey) -> Option<Cow<'repo, OsStr>>

Return the trusted string at key for launching using command::prepare(), or None if there is no such value or if no value was found in a trusted file.

impl Snapshot<'_>

Utilities and additional access

pub fn plumbing(&self) -> &File<'static>

Returns the underlying configuration implementation for a complete API, despite being a little less convenient.

It’s expected that more functionality will move up depending on demand.

impl Snapshot<'_>

pub fn credential_helpers( &self, url: Url, ) -> Result<(Cascade, Action, Options<'static>), Error>

Returns the configuration for all git-credential helpers from trusted configuration that apply to the given url along with an action preconfigured to invoke the cascade with. For details, please see this function.