gix::config

Struct SnapshotMut

pub struct SnapshotMut<'repo> {
    pub repo: Option<&'repo mut Repository>,
    /* private fields */
}

A platform to access configuration values and modify them in memory, while making them available when this platform is dropped as form of auto-commit. Note that the values will only affect this instance of the parent repository, and not other clones that may exist.

Note that these values won’t update even if the underlying file(s) change.

Use forget() to not apply any of the changes.

Fields
repo: Option<&'repo mut Repository>

The owning repository.

Implementations

impl<'repo> SnapshotMut<'repo>

Utilities

pub fn append_config( &mut self, values: impl IntoIterator<Item = impl AsRef<BStr>>, source: Source, ) -> Result<&mut Self, Error>

Append configuration values of the form core.abbrev=5 or remote.origin.url = foo or core.bool-implicit-true to the end of the repository configuration, with each section marked with the given source.

Note that doing so applies the configuration at the very end, so it will always override what came before it even though the source is of lower priority as what’s there.

pub fn commit(self) -> Result<&'repo mut Repository, Error>

Apply all changes made to this instance.

Note that this would also happen once this instance is dropped, but using this method may be more intuitive and won’t squelch errors in case the new configuration is partially invalid.

pub fn set_value<'b>( &mut self, key: &'static dyn Key, new_value: impl Into<&'b BStr>, ) -> Result<Option<BString>, Error>

Set the value at key to new_value, possibly creating the section if it doesn’t exist yet, or overriding the most recent existing value, which will be returned.

pub fn set_subsection_value<'a, 'b>( &mut self, key: &'static dyn Key, subsection: impl Into<&'a BStr>, new_value: impl Into<&'b BStr>, ) -> Result<Option<BString>, Error>

Set the value at key to new_value in the given subsection, possibly creating the section and sub-section if it doesn’t exist yet, or overriding the most recent existing value, which will be returned.

pub fn commit_auto_rollback(self) -> Result<CommitAutoRollback<'repo>, Error>

Create a structure the temporarily commits the changes, but rolls them back when dropped.

pub fn forget(self) -> File<'static>

Don’t apply any of the changes after consuming this instance, effectively forgetting them, returning the changed configuration.