gix::config

Module tree

The tree of supported configuration values for use in config_overrides or for validating and transforming well-known configuration values.

It can also be used to traverse all implemented keys and to validate values before usage as configuration overrides.

Leniency

When validating values, we don’t apply leniency here which is left to the caller. Leniency is an application defined configuration to ignore errors on non-security related values, which might make applications more resilient towards misconfiguration.

Modules

branch

checkout

core

credential

diff

extensions

fetch

gitoxide

http

index

key

keys

Generic value implementations for static instantiation.

protocol

push

remote

ssh

status

Structs

Author

The author top-level section.

Branch

The branch top-level section.

Checkout

The checkout top-level section.

Clone

The clone top-level section.

Committer

The committer top-level section.

Core

The core top-level section.

Credential

The credential top-level section.

Diff

The diff top-level section.

Extensions

The extension top-level section.

Fetch

The fetch top-level section.

Gitoxide

The gitoxide top-level section.

Http

The http top-level section.

Index

The index top-level section.

Init

The init top-level section.

Mailmap

Merge

Pack

The pack top-level section.

Protocol

The protocol top-level section.

Push

The push top-level section.

Remote

The remote top-level section.

Safe

The safe top-level section.

Ssh

The ssh top-level section.

Status

The status top-level section.

Url

The url top-level section.

User

The user top-level section.

Enums

Link

A way to link a key with other resources.

Note

A note attached to a key.

SubSectionRequirement

Determine how subsections may be used with a given key, suitable for obtaining the full name for use in assignments.

Traits

Key

A leaf-level entry in the git configuration, like url in remote.origin.url.

Section

Provide information about a configuration section.