gix::config::tree::gitoxide

Struct Core

pub struct Core;

The Core sub-section.

Implementations

impl Core

pub const DEFAULT_PACK_CACHE_MEMORY_LIMIT: UnsignedInteger

The gitoxide.core.defaultPackCacheMemoryLimit key.

pub const USE_NSEC: Boolean

The gitoxide.core.useNsec key.

pub const USE_STDEV: Boolean

The gitoxide.core.useStdev key.

pub const PROTECT_WINDOWS: Boolean

The gitoxide.core.protectWindows key.

pub const SHALLOW_FILE: Path

The gitoxide.core.shallowFile key.

pub const FILTER_PROCESS_DELAY: Boolean

The gitoxide.core.filterProcessDelay key (default true).

It controls whether or not long running filter driver processes can use the โdelayโ capability.

pub const EXTERNAL_COMMAND_STDERR: Boolean

The gitoxide.core.externalCommandStderr key (default true).

If true, the default, stderr of worktree filter programs, or any other git-context bearing command invoked will be inherited. If false, it will be suppressed completely.

pub const REFS_NAMESPACE: Any<RefsNamespace>

The gitoxide.core.refsNamespace key.