gix::config::tree::keys

Struct Any

pub struct Any<T: Validate = All> {
    pub name: &'static str,
    pub section: &'static dyn Section,
    pub subsection_requirement: Option<SubSectionRequirement>,
    pub link: Option<Link>,
    pub note: Option<Note>,
    /* private fields */
}

Implements a value without any constraints, i.e. a any value.

Fields
name: &'static str

The key of the value in the git configuration.

section: &'static dyn Section

The parent section of the key.

subsection_requirement: Option<SubSectionRequirement>

The subsection requirement to use.

link: Option<Link>

A link to other resources that might be eligible as value.

note: Option<Note>

A note about this key.

Implementations

impl Any<FullNameRef>

pub fn try_into_fullrefname( value: Cow<'_, BStr>, ) -> Result<Cow<'_, FullNameRef>, Error>

Return the validated full ref name from value if it is valid.

impl Any<Workers>

pub fn try_from_workers( &'static self, value: Result<i64, Error>, ) -> Result<usize, Error>

Return the amount of threads to use for checkout, with 0 meaning all available ones, after decoding our integer value from config, or None if the value isn’t set which is typically interpreted as “as many threads as available”

impl Any<CheckRoundTripEncoding>

pub fn try_into_encodings( &'static self, value: Option<Cow<'_, BStr>>, ) -> Result<Vec<&'static Encoding>, Error>

Convert value into a list of encodings, which are either space or coma separated. Fail if an encoding is unknown. If None, the default is returned.

impl Any<Eol>

pub fn try_into_eol( &'static self, value: Cow<'_, BStr>, ) -> Result<Mode, GenericErrorWithValue>

Convert value into the default end-of-line mode.

Deviation

git will allow any value and silently leaves it unset, we will fail if the value is not known.

impl Any<SafeCrlf>

pub fn try_into_safecrlf( &'static self, value: Cow<'_, BStr>, ) -> Result<CrlfRoundTripCheck, GenericErrorWithValue>

Convert value into the safe-crlf enumeration, if possible.

impl Any<AutoCrlf>

pub fn try_into_autocrlf( &'static self, value: Cow<'_, BStr>, ) -> Result<AutoCrlf, GenericErrorWithValue>

Convert value into the safe-crlf enumeration, if possible.

impl Any<LogAllRefUpdates>

pub fn try_into_ref_updates( &'static self, value: Option<Result<bool, Error>>, ) -> Result<Option<WriteReflog>, GenericErrorWithValue>

Returns the mode for ref-updates as parsed from value. If value is not a boolean, we try to interpret the string value instead. For correctness, this two step process is necessary as the interpretation of booleans in special in git-config, i.e. we can’t just treat it as string.

impl Any<CheckStat>

pub fn try_into_checkstat( &'static self, value: Cow<'_, BStr>, ) -> Result<bool, GenericErrorWithValue>

Returns true if the full set of stat entries should be checked, and it’s just as lenient as git.

impl Any<Abbrev>

pub fn try_into_abbreviation( &'static self, hex_len_str: Cow<'_, BStr>, object_hash: Kind, ) -> Result<Option<usize>, Error>

Convert the given hex_len_str into the amount of characters that a short hash should have. If None is returned, the correct value can be determined based on the amount of objects in the repo.

impl Any<Algorithm>

pub fn try_into_algorithm( &self, name: Cow<'_, BStr>, ) -> Result<Algorithm, Error>

Derive the diff algorithm identified by name, case-insensitively.

impl Any<Binary>

pub fn try_into_binary( &'static self, value: Option<Cow<'_, BStr>>, ) -> Result<Option<bool>, GenericErrorWithValue>

Convert value into a tri-state boolean that can take the special value auto, resulting in None, or is a boolean. If None is given, it’s treated as implicit boolean true, as this method is made to be used with gix_config::file::section::Body::value_implicit().

impl Any<Renames>

pub const fn new_renames( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

pub fn try_into_renames( &'static self, value: Result<bool, Error>, ) -> Result<Tracking, GenericError>

Try to convert the configuration into a valid rename tracking variant. Use value and if it’s an error, interpret the boolean as string

impl Any<ObjectFormat>

pub fn try_into_object_format( &'static self, value: Cow<'_, BStr>, ) -> Result<Kind, GenericErrorWithValue>

impl Any<NegotiationAlgorithm>

pub fn try_into_negotiation_algorithm( &'static self, name: Cow<'_, BStr>, ) -> Result<Algorithm, GenericErrorWithValue>

Derive the negotiation algorithm identified by name, case-sensitively.

impl Any<RecurseSubmodules>

pub fn try_into_recurse_submodules( &'static self, value: Result<bool, Error>, ) -> Result<FetchRecurse, GenericErrorWithValue>

Obtain the way submodules should be updated.

impl Any<RefsNamespace>

pub fn try_into_refs_namespace( &'static self, name: Cow<'_, BStr>, ) -> Result<Namespace, Error>

Derive the negotiation algorithm identified by name, case-sensitively.

impl Any<SslVersion>

pub const fn new_ssl_version( name: &'static str, section: &'static dyn Section, ) -> Self

impl Any<ProxyAuthMethod>

pub const fn new_proxy_auth_method( name: &'static str, section: &'static dyn Section, ) -> Self

impl Any<ExtraHeader>

pub fn try_into_extra_header( &'static self, values: Vec<Cow<'_, BStr>>, ) -> Result<Vec<String>, Error>

Convert a list of values into extra-headers, while failing entirely on illformed UTF-8.

impl Any<IndexThreads>

pub fn try_into_index_threads( &'static self, value: Cow<'_, BStr>, ) -> Result<usize, GenericErrorWithValue>

Parse value into the amount of threads to use, with 1 being single-threaded, or 0 indicating to select the amount of threads, with any other number being the specific amount of threads to use.

impl Any<IndexVersion>

pub fn try_into_index_version( &'static self, value: Result<i64, Error>, ) -> Result<Version, GenericError>

Try to interpret an integer value as index version.

impl Any<Allow>

pub fn try_into_allow( &'static self, value: Cow<'_, BStr>, scheme: Option<&str>, ) -> Result<Allow, Error>

Convert value into its respective Allow variant, possibly informing about the scheme we are looking at in the error.

impl Any<Version>

pub fn try_into_protocol_version( &'static self, value: Option<Result<i64, Error>>, ) -> Result<Protocol, GenericErrorWithValue>

Convert value into the corresponding protocol version, possibly applying the correct default.

impl Any<Default>

pub fn try_into_default( &'static self, value: Cow<'_, BStr>, ) -> Result<Default, GenericErrorWithValue>

Try to interpret value as push.default.

impl Any<TagOpt>

pub fn try_into_tag_opt( &'static self, value: Cow<'_, BStr>, ) -> Result<Tags, GenericErrorWithValue>

Try to interpret value as tag option.

Note

It’s heavily biased towards the git command-line unfortunately, and the only value of its kind. Maybe in future more values will be supported which are less about passing them to a sub-process.

impl Any<ShowUntrackedFiles>

pub fn try_into_show_untracked_files( &'static self, value: Cow<'_, BStr>, ) -> Result<UntrackedFiles, GenericErrorWithValue>

impl Any<DurationInMilliseconds>

pub const fn new_duration( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

pub fn try_into_duration( &'static self, value: Result<i64, Error>, ) -> Result<Duration, Error>

Return a valid duration as parsed from an integer that is interpreted as milliseconds.

impl Any<LockTimeout>

pub const fn new_lock_timeout( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

pub fn try_into_lock_timeout( &'static self, value: Result<i64, Error>, ) -> Result<Fail, Error>

Return information on how long to wait for locked files.

impl Any<PushRefSpec>

pub const fn new_push_refspec( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

impl Any<FetchRefSpec>

pub const fn new_fetch_refspec( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

impl Any<Url>

pub const fn new_url(name: &'static str, section: &'static dyn Section) -> Self

Create a new instance.

pub fn try_into_url(&'static self, value: Cow<'_, BStr>) -> Result<Url, Error>

Try to parse value as URL.

impl Any<UnsignedInteger>

pub const fn new_unsigned_integer( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

pub fn try_into_usize( &'static self, value: Result<i64, Error>, ) -> Result<usize, Error>

Convert value into a usize or wrap it into a specialized error.

pub fn try_into_u64( &'static self, value: Result<i64, Error>, ) -> Result<u64, Error>

Convert value into a u64 or wrap it into a specialized error.

pub fn try_into_u32( &'static self, value: Result<i64, Error>, ) -> Result<u32, Error>

Convert value into a u32 or wrap it into a specialized error.

impl Any<Time>

pub const fn new_time(name: &'static str, section: &'static dyn Section) -> Self

Create a new instance.

pub fn try_into_time( &self, value: Cow<'_, BStr>, now: Option<SystemTime>, ) -> Result<Time, Error>

Convert the value into a date if possible, with now as reference time for relative dates.

impl Any<Boolean>

pub const fn new_boolean( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

pub fn enrich_error( &'static self, value: Result<bool, Error>, ) -> Result<bool, Error>

Process the value into a result with an improved error message.

value is expected to be provided by gix_config::File::boolean().

impl Any<RemoteName>

pub const fn new_remote_name( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

pub fn try_into_symbolic_name( &'static self, name: Cow<'_, BStr>, ) -> Result<BString, Error>

Try to validate name as symbolic remote name and return it.

impl Any<All>

Init

pub const fn new(name: &'static str, section: &'static dyn Section) -> Self

Create a new instance from name and section

impl<T: Validate> Any<T>

Init other validate implementations

pub const fn new_with_validate( name: &'static str, section: &'static dyn Section, validate: T, ) -> Self

Create a new instance from name and section

impl<T: Validate> Any<T>

Builder

pub const fn with_subsection_requirement( self, requirement: Option<SubSectionRequirement>, ) -> Self

Set the subsection requirement to non-default values.

pub const fn with_environment_override(self, var: &'static str) -> Self

Associate an environment variable with this key.

This is mainly useful for enriching error messages.

pub const fn with_fallback(self, key: &'static dyn Key) -> Self

Set a link to another key which serves as fallback to provide a value if this key is not set.

pub const fn with_note(self, message: &'static str) -> Self

Attach an informative message to this key.

pub const fn with_deviation(self, message: &'static str) -> Self

Inform about a deviation in how this key is interpreted.

impl<T: Validate> Any<T>

Conversion

pub fn try_into_refspec( &'static self, value: Cow<'_, BStr>, op: Operation, ) -> Result<RefSpec, Error>

Try to convert value into a refspec suitable for the op operation.

pub fn try_into_string( &'static self, value: Cow<'_, BStr>, ) -> Result<String, Error>

Try to interpret value as UTF-8 encoded string.

impl Any<String>

pub const fn new_string( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

impl Any<Program>

pub const fn new_program( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

impl Any<Executable>

pub const fn new_executable( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.

impl Any<Path>

pub const fn new_path(name: &'static str, section: &'static dyn Section) -> Self

Create a new instance.