gix::config::tree::keys

Type Alias String

pub type String = Any<String>;

A key that represents a UTF-8 string.

Aliased Type

struct String {
    pub name: &'static str,
    pub section: &'static dyn Section,
    pub subsection_requirement: Option<SubSectionRequirement>,
    pub link: Option<Link>,
    pub note: Option<Note>,
    /* private fields */
}

Fields
name: &'static str

The key of the value in the git configuration.

section: &'static dyn Section

The parent section of the key.

subsection_requirement: Option<SubSectionRequirement>

The subsection requirement to use.

link: Option<Link>

A link to other resources that might be eligible as value.

note: Option<Note>

A note about this key.

Implementations

impl String

pub const fn new_string( name: &'static str, section: &'static dyn Section, ) -> Self

Create a new instance.