gix::config::tree::keys

Module validate

various implementations of the Validate trait.

Structs

All

Everything is valid.

Boolean

Assure that values that parse as git booleans are valid.

DurationInMilliseconds

Durations in milliseconds.

Executable

Values that are programs executables, everything is allowed.

FetchRefSpec

Values that parse as ref-specs for pushing.

LockTimeout

Timeouts used for file locks.

Path

Any path - everything is allowed.

Program

Values that are programs - everything is allowed.

PushRefSpec

Values that parse as ref-specs for pushing.

RemoteName

Values that are git remotes, symbolic or urls

String

A UTF-8 string.

Time

Assure that values that parse as git dates are valid.

UnsignedInteger

Assure that values that parse as unsigned integers are valid.

Url

Values that parse as URLs.