gix::config::tree

Trait Section

pub trait Section {
    // Required methods
    fn name(&self) -> &str;
    fn keys(&self) -> &[&dyn Key];

    // Provided methods
    fn sub_sections(&self) -> &[&dyn Section] { ... }
    fn parent(&self) -> Option<&dyn Section> { ... }
}

Provide information about a configuration section.

Required Methods

fn name(&self) -> &str

The section name, like remote in remote.origin.url.

fn keys(&self) -> &[&dyn Key]

The keys directly underneath it for carrying configuration values.

Provided Methods

fn sub_sections(&self) -> &[&dyn Section]

The list of sub-section names, which may be empty if there are no statically known sub-sections.

fn parent(&self) -> Option<&dyn Section>

The parent section if this is a statically known sub-section.