gix

Module create

Structs

Options

Options for use in into();

Enums

Error

The error used in into().

Kind

The kind of repository to create.

Functions

into

Create a new .git repository of kind within the possibly non-existing directory and return its path. Note that this is a simple template-based initialization routine which should be accompanied with additional corrections to respect git configuration, which is accomplished by its callers that return a Repository.