gix::create

Function into

pub fn into(
    directory: impl Into<PathBuf>,
    kind: Kind,
    _: Options,
) -> Result<Path, Error>

Create a new .git repository of kind within the possibly non-existing directory and return its path. Note that this is a simple template-based initialization routine which should be accompanied with additional corrections to respect git configuration, which is accomplished by its callers that return a Repository.