gix::diff

Module blob

For using text diffs, please have a look at the imara-diff documentation, maintained by Pascal Kuthe.

Modules

intern

pipeline

platform

sink

sources

unified_diff

Facilities to produce the unified diff format.

Structs

DiffLineStats

Information about the diff performed to detect similarity.

Driver

A set of values to define how to diff something that is associated with it using git-attributes, relevant for regular files.

Pipeline

A conversion pipeline to take an object or path from what’s stored in git to what can be diffed, while following the guidance of git-attributes at the respective path to learn if diffing should happen or if the content is considered binary.

Platform

A utility for performing a diff of two blobs, including flexible conversions, conversion-caching acquisition of diff information. Note that this instance will not call external filters as their output can’t be known programmatically, but it allows to prepare their input if the caller wishes to perform this task.

UnifiedDiff

A Sink that creates a textual diff in the format typically output by git or gnu-diff if the -u option is used, and passes it in full to a consumer.

UnifiedDiffBuilder

A Sink that creates a textual diff in the format typically output by git or gnu-diff if the -u option is used

Enums

Algorithm

imara-diff supports multiple different algorithms for computing an edit sequence. These algorithms have different performance and all produce different output.

ResourceKind

A way to classify a resource suitable for diffing.

Traits

Sink

Trait for processing the edit-scripts computed with diff

Functions

diff

Computes an edit-script that transforms input.before into input.after using the specified algorithm The edit-script is passed to sink.process_change while it is produced.

diff_with_tokens

Computes an edit-script that transforms before into after using the specified algorithm The edit-script is passed to sink.process_change while it is produced.