gix::diff::blob

Module intern

Structs

InternedInput

Two lists of interned tokens that can be compared with the diff function.

Interner

An interner that allows for fast access of tokens produced by a TokenSource.

Token

A token represented as an interned integer.

Traits

TokenSource