gix::diff::blob

Module pipeline

Modules

convert_to_diffable

Structs

Options

Options for use in a Pipeline.

Outcome

The outcome returned by Pipeline::convert_to_diffable().

WorktreeRoots

A way to access roots for different kinds of resources that are possibly located and accessible in a worktree.

Enums

Data

Data as part of an Outcome.

Mode

The specific way to convert a resource.