gix::diff::blob

Module sources

Structs

ByteLines

A TokenSource that returns the lines of a byte slice as tokens. See byte_lines and byte_lines_with_terminator for details

Lines

A TokenSource that returns the lines of a str as tokens. See lines and lines_with_terminator for details

Functions

byte_lines

Returns a TokenSource that uses the lines in data as Tokens. The newline seperator (\r\n or \n) is included in the emitted tokens. This means that changing the newline seperator from \r\n to \n (or omitting it fully on the last line) is detected by diff.

byte_lines_with_terminator

Returns a TokenSource that uses the lines in data as Tokens. A lines is a continous subslice of data which does not contain \n (or \r\n). The newline seperator (\r\n or \n) is not included in the emitted tokens. This means that changing the newline seperator from \r\n to \n (or omitting it fully on the last line) is not detected by diff.

lines

Returns a TokenSource that uses the lines in data as Tokens. The newline seperator (\r\n or \n) is not included in the emitted tokens. This means that changing the newline seperator from \r\n to \n (or omitting it fully on the last line) is not detected by diff.

lines_with_terminator

Returns a TokenSource that uses the lines in data as Tokens. The newline seperator (\r\n or \n) is included in the emitted tokens. This means that changing the newline seperator from \r\n to \n (or omitting it fully on the last line) is detected by diff.