gix::diff::blob::sources

Function lines

pub fn lines(data: &str) -> Lines<'_, false> 

Returns a TokenSource that uses the lines in data as Tokens. The newline seperator (\r\n or \n) is not included in the emitted tokens. This means that changing the newline seperator from \r\n to \n (or omitting it fully on the last line) is not detected by diff.