gix::diff::blob

Struct UnifiedDiffBuilder

pub struct UnifiedDiffBuilder<'a, W, T>
where
    W: Write,
    T: Display,
{ /* private fields */ }

A Sink that creates a textual diff in the format typically output by git or gnu-diff if the -u option is used

Implementations

impl<'a, T> UnifiedDiffBuilder<'a, String, T>
where T: Display,

pub fn new(input: &'a InternedInput<T>) -> UnifiedDiffBuilder<'a, String, T>

Create a new UnifiedDiffBuilder for the given input, that will return a String.

impl<'a, W, T> UnifiedDiffBuilder<'a, W, T>
where W: Write, T: Display,

pub fn with_writer( input: &'a InternedInput<T>, writer: W, ) -> UnifiedDiffBuilder<'a, W, T>

Create a new UnifiedDiffBuilder for the given input, that will writes it output to the provided implementation of Write.