gix::diff

Crate command

Launch commands very similarly to Command, but with git specific capabilities and adjustments.

Modules

shebang

Structs

Context

Additional information that is relevant to spawned processes, which typically receive a wealth of contextual information when spawned from git.

Prepare

A structure to keep settings to use when invoking a command via spawn(), after creating it with prepare().

Functions

extract_interpreter

Parse the shebang (#!<path>) from the first line of executable, and return the shebang data when available.

prepare

Prepare cmd for spawning by configuring it with various builder methods.