gix::diff

Module index

Structs

RewriteOptions

Options to configure how rewrites are tracked as part of the index() call.

Enums

Action

What to do after a ChangeRef was passed ot the callback of index().

ChangeRef

Identify a change that would have to be applied to lhs to obtain rhs, as provided in index().

Error

The error returned by index().

Type Aliases

Change

The fully-owned version of ChangeRef.