gix::diff::object

Module commit

Modules

message

ref_iter

Structs

ExtraHeaders

An iterator over extra headers in owned and borrowed commits.

MessageRef

A parsed commit message that assumes a title separated from the body by two consecutive newlines.

SignedData

The raw commit data, parseable by CommitRef or Commit, which was fed into a program to produce a signature.

Constants

SIGNATURE_FIELD_NAME

The well-known field name for gpg signatures.