gix::diff::object

Function compute_stream_hash

pub fn compute_stream_hash(
    hash_kind: Kind,
    object_kind: Kind,
    stream: &mut dyn Read,
    stream_len: u64,
    progress: &mut dyn Progress,
    should_interrupt: &AtomicBool,
) -> Result<ObjectId, Error>

A function to compute a hash of kind hash_kind for an object of object_kind and its data read from stream which has to yield exactly stream_len bytes. Use progress to learn about progress in bytes processed and should_interrupt to be able to abort the operation if set to true.