gix::diff::object

Trait Find

pub trait Find {
    // Required method
    fn try_find<'a>(
        &self,
        id: &oid,
        buffer: &'a mut Vec<u8>,
    ) -> Result<Option<Data<'a>>, Box<dyn Error + Send + Sync>>;
}

Find an object in the object store.

Notes

Find effectively needs generic associated types to allow a trait for the returned object type. Until then, we will have to make due with explicit types and give them the potentially added features we want.

Required Methods

fn try_find<'a>( &self, id: &oid, buffer: &'a mut Vec<u8>, ) -> Result<Option<Data<'a>>, Box<dyn Error + Send + Sync>>

Find an object matching id in the database while placing its raw, possibly encoded data into buffer.

Returns Some object if it was present in the database, or the error that occurred during lookup or object retrieval.