gix::diff::object

Trait FindHeader

pub trait FindHeader {
    // Required method
    fn try_header(
        &self,
        id: &oid,
    ) -> Result<Option<Header>, Box<dyn Error + Send + Sync>>;
}

Find the header of an object in the object store.

Required Methods

fn try_header( &self, id: &oid, ) -> Result<Option<Header>, Box<dyn Error + Send + Sync>>

Find the header of the object matching id in the database.

Returns Some header if it was present, or the error that occurred during lookup.