gix::diff

Module rewrites

Contains a Tracker to detect rewrites.

Modules

tracker

Types related to the rename tracker for renames, rewrites and copies.

Structs

Copies

Under which circumstances we consider a file to be a copy.

Outcome

Information collected while handling rewrites of files which may be tracked.

Tracker

A type to retain state related to an ongoing tracking operation to retain sets of interesting changes of which some are retained to at a later stage compute the ones that seem to be renames or copies.

Enums

CopySource

Determine in which set of files to search for copies.