gix::diff::rewrites

Module tracker

Types related to the rename tracker for renames, rewrites and copies.

Deviation

Note that the algorithm implemented here is in many ways different from what git does.

it’s less sophisticated and doesn’t use any ranking of candidates. Instead, it picks the first possible match.

the set used for copy-detection is probably smaller by default.

Modules

emit

visit

A module with types used in the user-callback in Tracker::emit().

Enums

ChangeKind

The kind of a change.

Traits

Change

A trait providing all functionality to abstract over the concept of a change, as seen by the Tracker.