gix::diff

Module tree

Modules

recorder

Useful for use as delegate implementing Visit to keep track of all seen changes. Useful for debugging or printing primarily.

visit

Structs

Recorder

A Visit implementation to record every observed change and keep track of the changed paths.

State

The state required to run tree-diffs.

Enums

Error

The error returned by tree().

Traits

Visit

A trait to allow responding to a traversal designed to figure out the changes to turn tree A into tree B.