gix::discover

Function upwards_opts

pub fn upwards_opts(
    directory: &Path,
    _: Options<'_>,
) -> Result<(Path, Trust), Error>

Find the location of the git repository directly in directory or in any of its parent directories and provide an associated Trust level by looking at the git directory’s ownership, and control discovery using options.

Fail if no valid-looking git repository could be found.