gix

Enum ObjectId

pub enum ObjectId {
    Sha1([u8; 20]),
}

An owned hash identifying objects, most commonly Sha1

Variants

Sha1([u8; 20])

A SHA 1 hash digest

Implementations

impl ObjectId

Hash decoding

pub fn from_hex(buffer: &[u8]) -> Result<ObjectId, Error>

Create an instance from a buffer of 40 bytes encoded with hexadecimal notation.

Such a buffer can be obtained using oid::write_hex_to(buffer)

impl ObjectId

Access and conversion

pub fn kind(&self) -> Kind

Returns the kind of hash used in this instance.

pub fn as_slice(&self) -> &[u8]

Return the raw byte slice representing this hash.

pub fn as_mut_slice(&mut self) -> &mut [u8]

Return the raw mutable byte slice representing this hash.

pub const fn empty_blob(hash: Kind) -> ObjectId

The hash of an empty blob.

pub const fn empty_tree(hash: Kind) -> ObjectId

The hash of an empty tree.

pub const fn null(kind: Kind) -> ObjectId

Returns an instances whose bytes are all zero.

pub fn is_null(&self) -> bool

Returns true if this hash consists of all null bytes.

pub fn is_empty_blob(&self) -> bool

Returns true if this hash is equal to an empty blob.

pub fn is_empty_tree(&self) -> bool

Returns true if this hash is equal to an empty tree.

impl ObjectId

Lifecycle

pub fn from_bytes_or_panic(bytes: &[u8]) -> ObjectId

Convert bytes into an owned object Id or panic if the slice length doesn’t indicate a supported hash.

Use Self::try_from(bytes) for a fallible version.

impl ObjectId

Sha1 hash specific methods

This impl block contains no items.