gix

Module env

Utilities to handle program arguments and other values of interest.

Modules

collate

Utilities to collate errors of common operations into one error type.

Functions

agent

Returns the name of the agent for identification towards a remote server as statically known when compiling the crate. Suitable for both git servers and HTTP servers, and used unless configured otherwise.

args_os

Equivalent to std::env::args_os(), but with precomposed unicode on MacOS and other apple platforms. It does not change the input arguments on any other platform.

args_os_opt

Like args_os(), but with the precompose_unicode parameter akin to core.precomposeUnicode in the Git configuration.

os_str_to_bstring

Convert the given input into a BString, useful for usage in clap.