gix::env

Module collate

Utilities to collate errors of common operations into one error type.

This is useful as this type can present an API to answer common questions, like whether a network request seems to have failed spuriously or if the underlying repository seems to be corrupted. Error collation supports all operations, including opening the repository.

Usage

The caller may define a function that specifies the result type as Result<T, gix::env::collate::{operation}::Error> to collect errors into a well-known error type which provides an API for simple queries.

Modules

fetch