gix::index

Module extension

Modules

decode

end_of_index_entry

link

sparse

tree

untracked_cache

Structs

FsMonitor

The extension for keeping state on recent information provided by the filesystem monitor.

Iter

An iterator over the data of index extensions.

Link

The link extension to track a shared index.

Tree

A structure to associate object ids of a tree with sections in the index entries list.

UntrackedCache

The extension for untracked files.

Constants

MIN_SIZE

The size of the smallest possible extension, which is no more than a signature and a 0 indicating its size.

Type Aliases

Signature

The kind of index extension.