gix

Module interrupt

Process-global interrupt handling

This module contains facilities to globally request an interrupt, which will cause supporting computations to abort once it is observed. Such checks for interrupts are provided in custom implementations of various traits to transparently add interrupt support to methods who wouldn’t otherwise by injecting it. see Read.

Structs

Iter

A wrapper for an inner iterator which will check for interruptions on each iteration.

Read

A wrapper for implementers of std::io::Read or std::io::BufRead with interrupt support.

Statics

IS_INTERRUPTED

The flag behind all utility functions in this module.

Functions

is_triggered

Returns true if an interrupt is requested.

reset

Sets the interrupt request to false, thus allowing those checking for is_triggered() to proceed.

trigger

Trigger an interrupt, signalling to those checking for is_triggered() to stop what they are doing.