gix::object::blob::diff

Struct Platform

pub struct Platform<'a> {
    pub resource_cache: &'a mut Platform,
}

A platform to keep temporary information to perform line diffs on modified blobs.

Fields
resource_cache: &'a mut Platform

The cache holding diffable data related to our blobs.

Implementations

impl Platform<'_>

pub fn lines<FnH, E>( &mut self, process_hunk: FnH, ) -> Result<Outcome<'_>, Error<E>>
where FnH: FnMut(Change<'_, '_>) -> Result<(), E>, E: Error + Send + Sync + 'static,

Perform a diff on lines between the old and the new version of a blob, passing each hunk of lines to process_hunk. The diffing algorithm is determined by the diff.algorithm configuration, or individual diff drivers. Note that process_hunk is not called if one of the involved resources are binary, but that can be determined by introspecting the outcome.

pub fn line_counts(&mut self) -> Result<Option<Counter<()>>, Error>

Count the amount of removed and inserted lines efficiently. Note that nothing will happen if one of the inputs is binary, and None will be returned.