gix::object::find::existing

Type Alias Error

pub type Error = Error;

An object could not be found in the database, or an error occurred when trying to obtain it.

Aliased Type

enum Error {
    Find(Box<dyn Error + Send + Sync>),
    NotFound {
        oid: ObjectId,
    },
}

Variants

Find(Box<dyn Error + Send + Sync>)

NotFound

Fields

oid: ObjectId