gix::object

Module tree

Modules

diff

traverse

Structs

Entry

An entry in a Tree, similar to an entry in a directory.

EntryMode

The mode of items storable in a tree, similar to the file mode on a unix file system.

EntryRef

An entry within a tree

Enums

EntryKind

A discretized version of ideal and valid values for entry modes.