gix::object::tree

Module diff

Modules

change

for_each

stats

Structs

Platform

The diffing platform returned by Tree::changes().

Stats

Provide aggregated information of a diff between two trees.

Enums

Action

Returned by the for_each function to control flow.

Change

Represents any possible change in order to turn one tree into another.

ChangeDetached

Represents any possible change in order to turn one tree into another, with fully-owned data.