gix::object::tree::diff

Enum Change

pub enum Change<'a, 'old, 'new> {
    Addition {
        location: &'a BStr,
        entry_mode: EntryMode,
        relation: Option<Relation>,
        id: Id<'new>,
    },
    Deletion {
        location: &'a BStr,
        entry_mode: EntryMode,
        relation: Option<Relation>,
        id: Id<'old>,
    },
    Modification {
        location: &'a BStr,
        previous_entry_mode: EntryMode,
        previous_id: Id<'old>,
        entry_mode: EntryMode,
        id: Id<'new>,
    },
    Rewrite {
        source_location: &'a BStr,
        source_relation: Option<Relation>,
        source_entry_mode: EntryMode,
        source_id: Id<'old>,
        diff: Option<DiffLineStats>,
        entry_mode: EntryMode,
        location: &'a BStr,
        id: Id<'new>,
        relation: Option<Relation>,
        copy: bool,
    },
}

Represents any possible change in order to turn one tree into another.

Variants

Addition

An entry was added, like the addition of a file or directory.

Fields

location: &'a BStr

The location of the file or directory, if tracking was enabled.

It may be empty if neither file names nor file paths are tracked.

entry_mode: EntryMode

The mode of the added entry.

relation: Option<Relation>

Identifies a relationship between this instance and another one, making it easy to reconstruct the top-level of directory changes.

id: Id<'new>

The object id of the added entry.

Deletion

An entry was deleted, like the deletion of a file or directory.

Fields

location: &'a BStr

The location of the file or directory, if tracking was enabled.

Otherwise, this value is always an empty path.

entry_mode: EntryMode

The mode of the deleted entry.

relation: Option<Relation>

Identifies a relationship between this instance and another one, making it easy to reconstruct the top-level of directory changes.

id: Id<'old>

The object id of the deleted entry.

Modification

An entry was modified, e.g. changing the contents of a file adjusts its object id and turning a file into a symbolic link adjusts its mode.

Fields

location: &'a BStr

The location of the file or directory, if tracking was enabled.

It may be empty if neither file names nor file paths are tracked.

previous_entry_mode: EntryMode

The mode of the entry before the modification.

previous_id: Id<'old>

The object id of the entry before the modification.

entry_mode: EntryMode

The mode of the entry after the modification.

id: Id<'new>

The object id after the modification.

Rewrite

Entries are considered rewritten if they are not trees and they, according to some understanding of identity, were renamed or copied. In case of renames, this means they originally appeared as Deletion signalling their source as well as an Addition acting as destination.

In case of copies, the copy flag is true and typically represents a perfect copy of a source was made.

This variant can only be encountered if rewrite tracking is enabled.

Note that mode changes may have occurred as well, i.e. changes from executable to non-executable or vice-versa.

Fields

source_location: &'a BStr

The location of the source of the rename operation.

It may be empty if neither file names nor file paths are tracked.

source_relation: Option<Relation>

Identifies a relationship between the source and another source, making it easy to reconstruct the top-level of directory changes.

source_entry_mode: EntryMode

The mode of the entry before the rename.

source_id: Id<'old>

The object id of the entry before the rename.

Note that this is the same as id if we require the similarity to be 100%, but may be different otherwise.

diff: Option<DiffLineStats>

Information about the diff we performed to detect similarity and match the source_id with the current state at id. Itโs None if source_id is equal to id, as identity made an actual diff computation unnecessary.

entry_mode: EntryMode

The mode of the entry after the rename. It could differ but still be considered a rename as we are concerned only about content.

location: &'a BStr

The location of the destination file or directory, if tracking was enabled.

It may be empty if neither file names nor file paths are tracked.

id: Id<'new>

The object id after the rename.

relation: Option<Relation>

Identifies a relationship between this destination and another destination, making it easy to reconstruct the top-level of directory changes.

copy: bool

If true, this rewrite is created by copy, and source_id is pointing to its source. Otherwise, itโs a rename, and source_id points to a deleted object, as renames are tracked as deletions and additions of the same or similar content.

Implementations

impl Change<'_, '_, '_>

pub fn diff<'b>( &self, resource_cache: &'b mut Platform, ) -> Result<Platform<'b>, Error>

Produce a platform for performing a line-diff no matter whether the underlying Change is an addition, modification, deletion or rewrite. Use resource_cache to store the diffable data and possibly reuse previously stored data, usually obtained with Repository::diff_resource_cache(). Afterward the platform, which holds on to resource_cache, can be used to perform ready-made operations on the pre-set resources.

Warning about Memory Consumption

resource_cache only grows, so one should call gix_diff::blob::Platform::clear_resource_cache occasionally.

impl<'a, 'old, 'new> Change<'a, 'old, 'new>

pub fn from_change_ref( change: ChangeRef<'a>, old_repo: &'old Repository, new_repo: &'new Repository, ) -> Self

Convert change into this instance type, attaching the old_repo and new_repo to each side respectively. Note that both repos are typically the same.

impl Change<'_, '_, '_>

Lifecycle

pub fn detach(self) -> ChangeDetached

Detach the repository instance to obtain a fully-owned version

impl Change<'_, '_, '_>

pub fn id(&self) -> Id<'_>

Return the current ID of the change.

pub fn location(&self) -> &BStr

Return the location of this instance.

pub fn entry_mode(&self) -> EntryMode

Return the current mode of this instance.