gix::object::tree::diff

Struct Platform

pub struct Platform<'a, 'repo> { /* private fields */ }

The diffing platform returned by Tree::changes().

Implementations

impl<'old> Platform<'_, 'old>

Add the item to compare to.

pub fn for_each_to_obtain_tree<'new, E>( &mut self, other: &Tree<'new>, for_each: impl FnMut(Change<'_, 'old, 'new>) -> Result<Action, E>, ) -> Result<Option<Outcome>, Error>
where E: Into<Box<dyn Error + Sync + Send + 'static>>,

Call for_each repeatedly with all changes that are needed to convert the source of the diff to the tree to other.

other could also be created with the empty_tree() method to handle the first commit in a repository - it doesn’t have a parent, equivalent to compare ‘nothing’ to something.

pub fn for_each_to_obtain_tree_with_cache<'new, E>( &mut self, other: &Tree<'new>, resource_cache: &mut Platform, for_each: impl FnMut(Change<'_, 'old, 'new>) -> Result<Action, E>, ) -> Result<Option<Outcome>, Error>
where E: Into<Box<dyn Error + Sync + Send + 'static>>,

Like Self::for_each_to_obtain_tree(), but with a reusable resource_cache which is used to perform diffs fast.

Reusing it between multiple invocations saves a lot of IOps as it avoids the creation of a temporary resource_cache that triggers reading or checking for multiple gitattribute files. Note that it’s recommended to call gix_diff::blob::Platform::clear_resource_cache() between the calls to avoid runaway memory usage, as the cache isn’t limited.

Note that to do rename tracking like git does, one has to configure the resource_cache with a conversion pipeline that uses gix_diff::blob::pipeline::Mode::ToGit.

impl Platform<'_, '_>

pub fn options(&mut self, change_opts: impl FnOnce(&mut Options)) -> &mut Self

Adjust diff options with change_opts.

impl Platform<'_, '_>

Convenience

pub fn stats(&mut self, other: &Tree<'_>) -> Result<Stats, Error>

Calculate statistics about the lines of the diff between our current and the other tree.

Performance Notes

Be sure to forcefully disable track_rewrites(None) to avoid rename tracking, an operation that doesn’t affect the statistics currently. As diffed resources aren’t cached, if highly repetitive blobs are expected, performance may be diminished. In real-world scenarios where blobs are mostly unique, that’s not an issue though.