gix

Module open

Modules

permissions

Various permissions to define what can be done when operating a Repository.

Structs

Options

The options used in ThreadSafeRepository::open_opts().

Permissions

Permissions associated with various resources of a git repository

Enums

Error

The error returned by crate::open().