gix::parallel

Function num_threads

pub fn num_threads(thread_limit: Option<usize>) -> usize

Returns the amount of threads the system can effectively use as the amount of its logical cores.

Only available with the parallel feature toggle set.