gix::parallel

Module reduce

Structs

IdentityWithResult

An identity reducer for those who want to use Stepwise or in_parallel() without the use of non-threaded reduction of products created in threads.

Stepwise

An iterator adaptor to allow running computations using in_parallel() in a step-wise manner, see the module docs for details.

Traits

Finalize

A trait reflecting the finalize() method of Reduce implementations

Reduce

An trait for aggregating items commonly produced in threads into a single result, without itself needing to be thread safe.