gix::path

Module env

Information about the environment in terms of locations of resources.

Functions

core_dir

Return the directory obtained by calling git --exec-path.

exe_invocation

Return the name of the Git executable to invoke it.

home_dir

Tries to obtain the home directory from HOME on all platforms, but falls back to [home::home_dir()] for more complex ways of obtaining a home directory, particularly useful on Windows.

installation_config

Return the location at which installation specific git configuration file can be found, or None if the binary could not be executed or its results could not be parsed.

installation_config_prefix

Return the location at which git installation specific configuration files are located, or None if the binary could not be executed or its results could not be parsed.

shell

Return the shell that Git would use, the shell to execute commands from.

system_prefix

Returns the platform dependent system prefix or None if it cannot be found (right now only on Windows).

var

Returns the contents of an environment variable of name with some special handling for certain environment variables (like HOME) for platform compatibility.

xdg_config

Returns the fully qualified path in the xdg-home directory (or equivalent in the home dir) to file, accessing env_var(<name>) to learn where these bases are.