gix::path::env

Function installation_config

pub fn installation_config() -> Option<&'static Path>

Return the location at which installation specific git configuration file can be found, or None if the binary could not be executed or its results could not be parsed.

Performance

This invokes the git binary which is slow on windows.