gix::path::env

Function shell

pub fn shell() -> &'static OsStr

Return the shell that Git would use, the shell to execute commands from.

On Windows, this is the full path to sh.exe bundled with Git for Windows if we can find it. If the bundled shell on Windows cannot be found, sh.exe is returned as the name of a shell, as it could possibly be found in PATH. On Unix it’s /bin/sh as the POSIX-compatible shell.

Note that the returned path might not be a path on disk, if it is a fallback path or if the file was moved or deleted since the first time this function is called.