gix::path

Function realpath

pub fn realpath(path: impl AsRef<Path>) -> Result<PathBuf, Error>

Check each component of path and see if it is a symlink. If so, resolve it. Do not fail for non-existing components, but assume these are as is.

If path is relative, the current working directory be used to make it absolute. Note that the returned path will be verbatim, and repositories with core.precomposeUnicode set will probably want to precompose the paths unicode.