gix::path

Function realpath_opts

pub fn realpath_opts(
    path: &Path,
    cwd: &Path,
    max_symlinks: u8,
) -> Result<PathBuf, Error>

The same as realpath(), but allow to configure max_symlinks to configure how many symbolic links we are going to follow. This serves to avoid running into cycles or doing unreasonable amounts of work.