gix

Module pathspec

Pathspec plumbing and abstractions

Modules

attributes

Parse .gitattribute files and provide utilities to match against them.

defaults

init

normalize

parse

search

Structs

Defaults

Default settings for some fields of a Pattern.

MagicSignature

Flags to represent ‘magic signatures’ which are parsed behind colons, like :top:.

Pattern

The output of a pathspec parsing operation. It can be used to match against a one or more paths.

Search

A lists of pathspec patterns, possibly from a file.

Enums

SearchMode

Parts of magic signatures which don’t stack as they all configure the way path specs are matched.

Functions

parse

Parse a git-style pathspec into a Pattern, setting the given default values in case these aren’t specified in input.