gix::pathspec

Crate attributes

Parse .gitattribute files and provide utilities to match against them.

Feature Flags

Modules

glob

Provide glob Patterns for matching against paths or anything else.

name

parse

search

state

Structs

Assignment

Name an attribute and describe it’s assigned state.

AssignmentRef

Holds validated attribute data as a reference

Name

Represents a validated attribute name

NameRef

Holds a validated attribute name as a reference

Search

A grouping of lists of patterns while possibly keeping associated to their base path in order to find matches.

Enums

Source

A list of known global sources for git attribute files in order of ascending precedence.

State

The state an attribute can be in, owning the value.

StateRef

The state an attribute can be in, referencing the value.

Functions

parse

Parse attribute assignments line by line from bytes, and fail the operation on error.