gix::pathspec::attributes

Module search

Structs

AttributeId

A type to denote an id of an attribute assignment for uniquely identifying each attribute or assignment.

Attributes

An implementation of the Pattern trait for attributes.

Match

Describes a matching pattern with

MatchLocation

Describes in which what file and line the match was found.

Metadata

Metadata associated with an attribute or macro name.

MetadataCollection

A utility type to collect metadata for each attribute, unified by its name.

Outcome

The result of a search, containing all matching attributes.

TrackedAssignment

A way to have an assignment (attr=value) but also associated it with an id that allows perfect mapping to tracking information. Note that the order is produced after the files are parsed as global ordering is needed that goes beyond the scope of a single Search instance.

Enums

MatchKind

The kind of attribute within the context of a match.

Value

A value of a pattern mapping, which is either a macro definition or a set of attributes.

Type Aliases

Assignments

A typically sized list of attributes.