gix

Module prelude

Traits

Finalize

A trait reflecting the finalize() method of Reduce implementations

Find

Find an object in the object store.

FindExt

An extension trait with convenience functions.

Header

A way to obtain object properties without fully decoding it.

HeaderExt

An extension trait with convenience functions.

ObjectIdExt

An extension trait to add functionality to ObjectIds.

ReferenceExt

Extensions for [references][gix_ref::Reference].

TreeDiffChangeExt

Extensions for Change.

TreeEntryExt

Extensions for Entry.

TreeEntryRefExt

Extensions for EntryRef.

TreeIterExt

An extension trait for tree iterators

Write

Describe the capability to write git objects into an object store.