gix::prelude

Trait Header

pub trait Header {
    // Required method
    fn try_header(
        &self,
        id: &oid,
    ) -> Result<Option<Header>, Box<dyn Error + Send + Sync>>;
}

A way to obtain object properties without fully decoding it.

Required Methods

fn try_header( &self, id: &oid, ) -> Result<Option<Header>, Box<dyn Error + Send + Sync>>

Try to read the header of the object associated with id or return None if it could not be found.