gix::prelude

Trait TreeIterExt

pub trait TreeIterExt: Sealed {
    // Required method
    fn traverse<StateMut, Find, V>(
        &self,
        state: StateMut,
        objects: Find,
        delegate: &mut V,
    ) -> Result<(), Error>
       where Find: Find,
             StateMut: BorrowMut<State>,
             V: Visit;
}

An extension trait for tree iterators

Required Methods

fn traverse<StateMut, Find, V>( &self, state: StateMut, objects: Find, delegate: &mut V, ) -> Result<(), Error>
where Find: Find, StateMut: BorrowMut<State>, V: Visit,

Traverse this tree with state being provided to potentially reuse allocations, and find being a function to lookup trees and turn them into iterators.

The delegate implements a way to store details about the traversal to allow paying only for what’s actually used. Since it is expected to store the operation result, unit is returned.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.