gix::progress::prodash

Module messages

Structs

Message

A message to be stored along with the progress tree.

MessageCopyState

State used to keep track of what’s new since the last time message were copied.

MessageRingBuffer

A ring buffer for messages.

Enums

MessageLevel

The severity of a message