gix::progress::prodash::messages

Struct MessageRingBuffer

pub struct MessageRingBuffer { /* private fields */ }

A ring buffer for messages.

Implementations

impl MessageRingBuffer

pub fn with_capacity(capacity: usize) -> MessageRingBuffer

Create a new instance the ability to hold capacity amount of messages.

pub fn push_overwrite( &mut self, level: MessageLevel, origin: String, message: impl Into<String>, )

Push a message from origin at severity level into the buffer, possibly overwriting the last message added.

pub fn copy_all(&self, out: &mut Vec<Message>)

Copy all messages currently contained in the buffer to out.

pub fn copy_new( &self, out: &mut Vec<Message>, previous: Option<MessageCopyState>, ) -> MessageCopyState

Copy all new messages into out that where received since the last time this method was called provided its previous return value.