gix::progress::prodash

Module progress

Modules

key

Structs

Discard

An implementation of NestedProgress which discards all calls.

DoOrDiscard

An implementation of Progress which can be created easily from Option<impl Progress>.

Key

A type identifying a spot in the hierarchy of Tree items.

Task

The value associated with a spot in the hierarchy.

ThroughputOnDrop

Emit a message with throughput information when the instance is dropped.

Value

Progress associated with some item in the progress tree.

Enums

Either

An implementation of NestedProgress showing either one or the other implementation.

State

Indicate whether a progress can or cannot be made.

Constants

UNKNOWN

The default Id to use if there is no need for an id.

Type Aliases

AtomicStep

The amount of steps a progress can make, for threadsafe counting.

Id

Four bytes of function-local unique and stable identifier for each item added as progress, like b“TREE“ or b“FILE“.

Step

The amount of steps a progress can make

StepShared

As step, but shareable.