gix::progress::prodash::progress::key

Struct Adjacency

pub struct Adjacency(pub SiblingLocation, pub SiblingLocation, pub SiblingLocation, pub SiblingLocation, pub SiblingLocation, pub SiblingLocation);

A type providing information about whatโs above and below Tree items.

Tuple Fields
0: SiblingLocation1: SiblingLocation2: SiblingLocation3: SiblingLocation4: SiblingLocation5: SiblingLocation
Implementations

impl Adjacency

pub fn level(&self) -> u8

Return the level at which this sibling is located in the hierarchy.

pub fn get(&self, level: u8) -> Option<&SiblingLocation>

Get a reference to the sibling location at level.

pub fn get_mut(&mut self, level: u8) -> Option<&mut SiblingLocation>

Get a mutable reference to the sibling location at level.