gix::progress::prodash

Module unit

Modules

display

Various utilities to display values and units.

Structs

Range

A helper for formatting numbers representing ranges in renderers as in 2 of 5 steps.

Unit

A configurable and flexible unit for use in Progress::init().

Enums

Kind

Either a static label or a dynamic one implementing DisplayValue.

Traits

DisplayValue

A trait to encapsulate all capabilities needed to display a value with unit within a renderer.

Functions

dynamic

Returns a unit that is a dynamic label.

dynamic_and_mode

Returns a unit that is a dynamic label along with information on where to display a fraction and throughput.

label

Returns a unit that is a static label.

label_and_mode

Returns a unit that is a static label along with information on where to display a fraction and throughput.