gix::progress

Struct Read

pub struct Read<T, P> {
    pub inner: T,
    pub progress: P,
}

A structure passing every read call through to the contained Progress instance using inc_by(bytes_read).

Fields
inner: T

The implementor of std::io::Read to which progress is added

progress: P

The progress instance receiving progress information on each invocation of reader