gix::reference::iter

Struct Platform

pub struct Platform<'r> {
    pub repo: &'r Repository,
    /* private fields */
}

A platform to create iterators over references.

Fields
repo: &'r Repository

The owning repository.

Implementations

impl Platform<'_>

pub fn all(&self) -> Result<Iter<'_>, Error>

Return an iterator over all references in the repository.

Even broken or otherwise unparsable or inaccessible references are returned and have to be handled by the caller on a case by case basis.

pub fn prefixed<'a>( &self, prefix: impl TryInto<&'a RelativePath, Error = Error>, ) -> Result<Iter<'_>, Error>

Return an iterator over all references that match the given prefix.

These are of the form refs/heads/ or refs/remotes/origin, and must not contain relative paths components like . or ...

pub fn tags(&self) -> Result<Iter<'_>, Error>

Return an iterator over all references that are tags.

They are all prefixed with refs/tags.

pub fn local_branches(&self) -> Result<Iter<'_>, Error>

Return an iterator over all local branches.

They are all prefixed with refs/heads.

pub fn remote_branches(&self) -> Result<Iter<'_>, Error>

Return an iterator over all remote branches.

They are all prefixed with refs/remotes.