gix::reference::log

Function message

pub fn message(operation: &str, message: &BStr, num_parents: usize) -> BString

Generate a message typical for git commit logs based on the given operation, commit message and num_parents of the commit.