gix

Module remote

Modules

connect

fetch

find

init

name

ref_map

save

url

Structs

Connection

A type to represent an ongoing connection to a remote host, typically with the connection already established.

Enums

Direction

The direction of an operation carried out (or to be carried out) through a remote.

Name

The name of a remote, either interpreted as symbol like origin or as url as returned by Remote::name().

Type Aliases

AuthenticateFn

A function that performs a given credential action, trying to obtain credentials for an operation that needs it.

Names

A type-definition for a sorted list of unvalidated remote names - they have been read straight from the configuration.