gix::remote

Module fetch

Modules

negotiate

outcome

Additional types related to the outcome of a fetch operation.

prepare

refmap

refs

Structs

Outcome

The outcome of receiving a pack via Prepare::receive().

Prepare

A structure to hold the result of the handshake with the remote and configure the upcoming fetch operation.

RefMap

Information about the relationship between our refspecs, and remote references with their local counterparts.

Enums

Error

The error returned by receive().

ProgressId

The progress ids used in during various steps of the fetch operation.

RefLogMessage

The way reflog messages should be composed whenever a ref is written with recent objects from a remote.

Shallow

Describe how shallow clones are handled when fetching, with variants defining how the shallow boundary is handled.

Status

The status of the repository after the fetch operation

Tags

Describe how to handle tags when fetching