gix::remote

Type Alias AuthenticateFn

pub type AuthenticateFn<'a> = Box<dyn FnMut(Action) -> Result + 'a>;

A function that performs a given credential action, trying to obtain credentials for an operation that needs it.

Aliased Type

struct AuthenticateFn<'a>(/* private fields */);