gix::revision

Module walk

Modules

iter

Structs

Info

Information about a commit that we obtained naturally as part of the iteration.

Platform

A platform to traverse the revision graph by adding starting points as well as points which shouldn’t be crossed, returned by Repository::rev_walk().

Enums

Error

The error returned by Platform::all() and Platform::selected().

Sorting

Specify how to sort commits during a revision::Walk traversal.