gix

Module status

Re-exports

pub use gix_status as plumbing;

Modules

index_worktree

into_iter

is_dirty

iter

tree_index

Structs

Iter

An iterator for changes between the index and the worktree and the head-tree and the index.

Outcome

The data the thread sends over to the receiving iterator.

Platform

A structure to hold options configuring the status request, which can then be turned into an iterator.

Enums

Error

The error returned by status().

Item

The item produced by the iterator.

Submodule

How to obtain a submodule’s status.

UntrackedFiles

How untracked files should be handled.