gix::status

Module index_worktree

Modules

iter

Structs

BuiltinSubmoduleStatus

An implementation of a trait to use with Repository::index_worktree_status() to compute the submodule status using Submodule::status().

Iter

An iterator for changes between the index and the worktree.

Options

Options for use with Repository::index_worktree_status().

Enums

Error

The error returned by Repository::index_worktree_status().

Item

The item produced by the iterator

RewriteSource

Either an index entry for renames or another directory entry in case of copies.