gix::status::index_worktree

Enum Item

pub enum Item {
    Modification {
        entry: Entry,
        entry_index: usize,
        rela_path: BString,
        status: EntryStatus<(), Status>,
    },
    DirectoryContents {
        entry: Entry,
        collapsed_directory_status: Option<Status>,
    },
    Rewrite {
        source: RewriteSource,
        dirwalk_entry: Entry,
        dirwalk_entry_collapsed_directory_status: Option<Status>,
        dirwalk_entry_id: ObjectId,
        diff: Option<DiffLineStats>,
        copy: bool,
    },
}

The item produced by the iterator

Variants

Modification

A tracked file was modified, and index-specific information is passed.

Fields

entry: Entry

The entry with modifications.

entry_index: usize

The index of the entry for lookup in gix_index::State::entries() - useful to look at neighbors.

rela_path: BString

The repository-relative path of the entry.

status: EntryStatus<(), Status>

The computed status of the entry.

DirectoryContents

An entry returned by the directory walk, without any relation to the index.

This can happen if ignored files are returned as well, or if rename-tracking is disabled.

Fields

entry: Entry

The entry found during the disk traversal.

collapsed_directory_status: Option<Status>

collapsed_directory_status is Some(dir_status) if this entry was part of a directory with the given dir_status that wasn’t the same as the one of entry and if [gix_dir::walk::Options::emit_collapsed] was CollapsedEntriesEmissionMode::OnStatusMismatch. It will also be Some(dir_status) if that option was CollapsedEntriesEmissionMode::All.

Rewrite

The rewrite tracking discovered a match between a deleted and added file, and considers them equal enough, depending on the tracker settings.

Note that the source of the rewrite is always the index as it detects the absence of entries, something that can’t be done during a directory walk.

Fields

source: RewriteSource

The source of the rewrite operation.

dirwalk_entry: Entry

The untracked entry found during the disk traversal, the destination of the rewrite.

Note that its rela_path is the destination of the rewrite, and the current location of the entry.

dirwalk_entry_collapsed_directory_status: Option<Status>

collapsed_directory_status is Some(dir_status) if this dirwalk_entry was part of a directory with the given dir_status that wasn’t the same as the one of dirwalk_entry and if [gix_dir::walk::Options::emit_collapsed] was CollapsedEntriesEmissionMode::OnStatusMismatch. It will also be Some(dir_status) if that option was CollapsedEntriesEmissionMode::All.

dirwalk_entry_id: ObjectId

The object id as it would appear if the entry was written to the object database, specifically hashed in order to determine equality. Note that it doesn’t (necessarily) exist in the object database, and may be null if no hashing was performed.

diff: Option<DiffLineStats>

It’s None if the ‘source.id’ is equal to dirwalk_entry_id, as identity made an actual diff computation unnecessary. Otherwise, and if enabled, it’s Some(stats) to indicate how similar both entries were.

copy: bool

If true, this rewrite is created by copy, and ‘source.id’ is pointing to its source. Otherwise, it’s a rename, and ‘source.id’ points to a deleted object, as renames are tracked as deletions and additions of the same or similar content.

Implementations

impl Item

pub fn summary(&self) -> Option<Summary>

Return a simplified summary of the item as digest of its status, or None if this item is created from the directory walk and is not untracked, or if it is merely to communicate a needed update to the index entry.

pub fn rela_path(&self) -> &BStr

The repository-relative path of the entry contained in this item.