gix::status::index_worktree

Struct Options

pub struct Options {
    pub sorting: Option<Sorting>,
    pub dirwalk_options: Option<Options>,
    pub rewrites: Option<Rewrites>,
    pub thread_limit: Option<usize>,
}

Options for use with Repository::index_worktree_status().

Fields
sorting: Option<Sorting>

The way all output should be sorted.

If None, and depending on the rewrites field, output will be immediate but the output order isnโt determined, and may differ between two runs. rewrites also depend on the order of entries that are presented to it, hence for deterministic results, sorting needs to be enabled.

If Some(_), all entries are collected beforehand, so they can be sorted before outputting any of them to the user.

If immediate output of entries in any order is desired, this should be None, along with rewrites being None as well.

dirwalk_options: Option<Options>

If not None, the options to configure the directory walk, determining how its results will look like.

If None, only modification checks are performed.

Can be instantiated with Repository::dirwalk_options().

rewrites: Option<Rewrites>

If Some(_), along with Some(_) in dirwalk_options, rewrite tracking will be performed between the index and the working tree. Note that there is no git-configuration specific to index-worktree rename tracking. When rewrite tracking is enabled, there will be a delay for some entries as they partake in the rename-analysis.

thread_limit: Option<usize>

If set, donโt use more than this amount of threads for the tracked modification check. Otherwise, usually use as many threads as there are logical cores. A value of 0 is interpreted as no-limit