gix::status

Struct Outcome

pub struct Outcome {
    pub index_worktree: Outcome,
    pub tree_index: Option<Outcome>,
    pub worktree_index: IndexPersistedOrInMemory,
    /* private fields */
}

The data the thread sends over to the receiving iterator.

Fields
index_worktree: Outcome

The outcome of the index-to-worktree comparison operation.

tree_index: Option<Outcome>

The outcome of the diff between HEAD^{tree} and the index, or None if this outcome was produced with the into_index_worktree_iter().

worktree_index: IndexPersistedOrInMemory

The worktree index that was used for the operation.

Implementations

impl Outcome

pub fn has_changes(&self) -> bool

Returns true if the index has received currently unapplied changes that should be written back.

If they are not written back, subsequent status operations will take longer to complete, whereas the additional work can be prevented by writing the changes back to the index.

pub fn write_changes(&mut self) -> Option<Result<(), Error>>

Write the changes if there are any back to the index file. This can only be done once as the changes are consumed in the process, if there were any.