gix::status

Struct Platform

pub struct Platform<'repo, Progress>
where
    Progress: Progress + 'static,
{ /* private fields */ }

A structure to hold options configuring the status request, which can then be turned into an iterator.

Implementations

impl<Progress> Platform<'_, Progress>
where Progress: Progress,

Builder

pub fn dirwalk_options(self, cb: impl FnOnce(Options) -> Options) -> Self

Call cb on dirwalk options if these are set (which is the default when created through Repository::status()). The directory walk is used to find untracked files or ignored files.

cb will be able to run builder-methods on the passed dirwalk options.

pub fn dirwalk_options_mut( &mut self, cb: impl FnOnce(&mut Options), ) -> &mut Self

Like dirwalk_options(), but taking a mutable instance instead.

pub fn untracked_files(self, untracked_files: UntrackedFiles) -> Self

A simple way to explicitly set the desired way of listing untracked_files, overriding any value set by the git configuration.

Note that if None is used, the directory walk will be disabled entirely after this call. Further, if no dirwalk options are present anymore, this call has no effect.

pub fn should_interrupt_shared( self, should_interrupt: &'static AtomicBool, ) -> Self

Set the interrupt flag to should_interrupt, which typically is an application-wide flag that is ultimately controlled by user interrupts.

If it is true, the iteration will stop immediately.

pub fn should_interrupt_owned(self, should_interrupt: Arc<AtomicBool>) -> Self

Set the interrupt flag to should_interrupt, as controlled by the caller.

If it is true, the iteration will stop immediately.

pub fn index_worktree_submodules( self, submodules: impl Into<Option<Submodule>>, ) -> Self

Configure how the submodule_status is obtained when looking at submodules that are still mentioned in the index.

pub fn index(self, index: IndexPersistedOrInMemory) -> Self

Set the index to use when making comparisons to the worktree and the head revision.

Defaults to the current index, or an empty one if it doesn’t exist (yet).

pub fn index_worktree_rewrites( self, rewrites: impl Into<Option<Rewrites>>, ) -> Self

Configure the index-to-worktree rename tracking with rewrites, which is None by default.

Note that Git does not have configuration related to rename tracking of changes between the index and the worktree. The closest there is can be obtained using crate::diff::new_rewrites(), which refers to rename tracking between trees.

Also note that if rewrites are Some(), sorting will automatically be configured to assure deterministic outcomes for rewrite solutions.

pub fn index_worktree_options_mut(self, cb: impl FnOnce(&mut Options)) -> Self

Adjust all options related to the index-worktree status. This is a catch-all in case there are no more specific methods that could be used instead to change the respective option.

pub fn head_tree(self, tree: impl Into<ObjectId>) -> Self

Set the tree at which the HEAD ref ought to reside. Setting this explicitly allows to compare the index to a tree that it possibly didn’t originate from.

If not set explicitly, it will be read via HEAD^{tree}.

pub fn tree_index_track_renames(self, renames: TrackRenames) -> Self

Configure how rename tracking should be performed when looking at changes between the head tree and the index.

impl<Progress> Platform<'_, Progress>
where Progress: Progress,

Lifecycle

pub fn into_index_worktree_iter( self, patterns: impl IntoIterator<Item = BString>, ) -> Result<Iter, Error>

Turn the platform into an iterator for changes between the index and the working tree.

patterns

Optional patterns to use to limit the paths to look at. If empty, all paths are considered.

impl<Progress> Platform<'_, Progress>
where Progress: Progress,

Lifecycle

pub fn into_iter( self, patterns: impl IntoIterator<Item = BString>, ) -> Result<Iter, Error>

Turn the platform into an iterator for changes between the head-tree and the index, and the index and the working tree, while optionally listing untracked and/or ignored files.

patterns

Optional patterns to use to limit the paths to look at. If empty, all paths are considered.

Important: Undefined ordering

When compiled with the parallel feature, three operations are run at once:

a dirwalk to find untracked and possibly ignored files

an entry-by-entry check to see which of the tracked files changed, and how

a tree-index comparison

All of these generate distinct events which may now happen in any order, so consumers that are ordering dependent have to restore their desired order.

This isn’t feasible to do here as it would mean that returned items would have to be delayed, degrading performance for everyone who isn’t order-dependent.