gix

Struct AttributeStack

pub struct AttributeStack<'repo> {
    pub repo: &'repo Repository,
    /* private fields */
}

A utility to access .gitattributes and .gitignore information efficiently.

Fields
repo: &'repo Repository

The owning repository.

Implementations

impl<'repo> AttributeStack<'repo>

Lifecycle

pub fn new(stack: Stack, repo: &'repo Repository) -> Self

Create a new instance from a repo and the underlying pre-configured stack.

Note that this type is typically created by Repository::attributes() or Repository::attributes_only().

pub fn detach(self) -> Stack

Detach the repository and return the underlying plumbing datatype.

impl AttributeStack<'_>

Platform retrieval

pub fn at_path( &mut self, relative: impl AsRef<Path>, mode: Option<Mode>, ) -> Result<Platform<'_>>

Append the relative path to the root directory of the cache and load all attribute or ignore files on the way as needed. Use mode to specify what kind of item lives at relative - directories may match against rules specifically. If mode is None, the item at relative is assumed to be a file.

The returned platform may be used to access the actual attribute or ignore information.

pub fn at_entry( &mut self, relative: impl ToNormalPathComponents, mode: Option<Mode>, ) -> Result<Platform<'_>>

Obtain a platform for attribute or ignore lookups from a repo-relative path, typically obtained from an index entry. mode should reflect whether itโs a directory or not, or left at None if unknown.

If relative ends with / and mode is None, it is automatically assumed to be a directory.