gix

Struct Commit

pub struct Commit<'repo> {
    pub id: ObjectId,
    pub data: Vec<u8>,
    pub repo: &'repo Repository,
}

A decoded commit object with access to its owning repository.

Fields
id: ObjectId

The id of the commit

data: Vec<u8>

The fully decoded commit data

repo: &'repo Repository

The owning repository.

Implementations

impl Commit<'_>

Remove Lifetime

pub fn detached(&self) -> ObjectDetached

Create an owned instance of this object, copying our data in the process.

pub fn detach(self) -> ObjectDetached

Sever the connection to the Repository and turn this instance into a standalone object.

pub fn take_data(&mut self) -> Vec<u8>

Retrieve this instanceโs encoded data, leaving its own data empty.

This method works around the immovability of members of this type.

impl<'repo> Commit<'repo>

pub fn short_id(&self) -> Result<Prefix, Error>

Turn this objects id into a shortened id with a length in hex as configured by core.abbrev.

pub fn message(&self) -> Result<MessageRef<'_>, Error>

Parse the commits message into a MessageRef

pub fn message_raw(&self) -> Result<&BStr, Error>

Decode the commit object until the message and return it.

pub fn message_raw_sloppy(&self) -> &BStr

Obtain the message by using intricate knowledge about the encoding, which is fastest and canโt fail at the expense of error handling.

pub fn time(&self) -> Result<Time, Error>

Decode the commit and obtain the time at which the commit was created.

For the time at which it was authored, refer to .decode()?.author.time.

pub fn decode(&self) -> Result<CommitRef<'_>, Error>

Decode the entire commit object and return it for accessing all commit information.

It will allocate only if there are more than 2 parents.

Note that the returned commit object does make lookup easy and should be used for successive calls to string-ish information to avoid decoding the object more than once.

pub fn iter(&self) -> CommitRefIter<'_>

Return an iterator over tokens, representing this commit piece by piece.

pub fn author(&self) -> Result<SignatureRef<'_>, Error>

Return the commits author, with surrounding whitespace trimmed.

pub fn committer(&self) -> Result<SignatureRef<'_>, Error>

Return the commits committer. with surrounding whitespace trimmed.

pub fn parent_ids(&self) -> impl Iterator<Item = Id<'repo>> + '_

Decode this commits parent ids on the fly without allocating.

pub fn tree(&self) -> Result<Tree<'repo>, Error>

Parse the commit and return the tree object it points to.

pub fn tree_id(&self) -> Result<Id<'repo>, Error>

Parse the commit and return the tree id it points to.

pub fn id(&self) -> Id<'repo>

Return our id own id with connection to this repository.

pub fn ancestors(&self) -> Platform<'repo>

Obtain a platform for traversing ancestors of this commit.

pub fn signature( &self, ) -> Result<Option<(Cow<'_, BStr>, SignedData<'_>)>, Error>

Extracts the PGP signature and the data that was used to create the signature, or None if it wasnโt signed.