gix

Struct Tag

pub struct Tag<'repo> {
    pub id: ObjectId,
    pub data: Vec<u8>,
    pub repo: &'repo Repository,
}

A decoded tag object with access to its owning repository.

Fields
id: ObjectId

The id of the tree

data: Vec<u8>

The fully decoded tag data

repo: &'repo Repository

The owning repository.

Implementations

impl<'repo> Tag<'repo>

pub fn decode(&self) -> Result<TagRef<'_>, Error>

Decode the entire tag object and return it for accessing all tag information.

This never allocates.

Note that the returned commit object does make lookup easy and should be used for successive calls to string-ish information to avoid decoding the object more than once.

pub fn target_id(&self) -> Result<Id<'repo>, Error>

Decode this tag partially and return the id of its target.

pub fn tagger(&self) -> Result<Option<SignatureRef<'_>>, Error>

Decode this tag partially and return the tagger, if the field exists.

impl Tag<'_>

Remove Lifetime

pub fn detached(&self) -> ObjectDetached

Create an owned instance of this object, copying our data in the process.

pub fn detach(self) -> ObjectDetached

Sever the connection to the Repository and turn this instance into a standalone object.

pub fn take_data(&mut self) -> Vec<u8>

Retrieve this instanceโs encoded data, leaving its own data empty.

This method works around the immovability of members of this type.