gix

Struct oid

pub struct oid { /* private fields */ }

A borrowed reference to a hash identifying objects.

Future Proofing

In case we wish to support multiple hashes with the same length we cannot discriminate using the slice length anymore. To make that work, we will use the high bits of the internal bytes slice length (a fat pointer, pointing to data and its length in bytes) to encode additional information. Before accessing or returning the bytes, a new adjusted slice will be constructed, while the high bits will be used to help resolving the hash kind(). We expect to have quite a few bits available for such ‘conflict resolution’ as most hashes aren’t longer than 64 bytes.

Implementations

impl oid

Conversion

pub fn try_from_bytes(digest: &[u8]) -> Result<&oid, Error>

Try to create a shared object id from a slice of bytes representing a hash digest

pub fn from_bytes_unchecked(value: &[u8]) -> &oid

Create an OID from the input value slice without performing any safety check. Use only once sure that value is a hash of valid length.

impl oid

Access

pub fn kind(&self) -> Kind

The kind of hash used for this instance.

pub fn first_byte(&self) -> u8

The first byte of the hash, commonly used to partition a set of object ids.

pub fn as_bytes(&self) -> &[u8]

Interpret this object id as raw byte slice.

pub fn to_hex_with_len(&self, len: usize) -> HexDisplay<'_>

Return a type which can display itself in hexadecimal form with the len amount of characters.

pub fn to_hex(&self) -> HexDisplay<'_>

Return a type which displays this oid as hex in full.

pub fn is_null(&self) -> bool

Returns true if this hash consists of all null bytes.

impl oid

Sha1 specific methods

pub fn hex_to_buf<'a>(&self, buf: &'a mut [u8]) -> &'a mut str

Write ourselves to the out in hexadecimal notation, returning the hex-string ready for display.

Panics if the buffer isn’t big enough to hold twice as many bytes as the current binary size.

pub fn write_hex_to(&self, out: &mut dyn Write) -> Result<(), Error>

Write ourselves to out in hexadecimal notation.

impl oid

pub fn verify(&self, expected: &oid) -> Result<(), Error>

Verify that self matches the expected object ID.

Returns an Error containing both object IDs if they differ.