gix

Module submodule

Submodule plumbing and abstractions

Modules

config

fetch_recurse

head_id

index_id

is_active

is_active_platform

modules

open

open_modules_file

status

Structs

File

All relevant information about a git module, typically from .gitmodules files.

IsActivePlatform

A platform to keep the state necessary to perform repeated active checks, created by File::is_active_platform().

State

A summary of the state of all parts forming a submodule, which allows to answer various questions about it.

Status

A simplified status of the Submodule.

Type Aliases

ModulesSnapshot

A lazily loaded and auto-updated worktree index.