gix

Type Alias RefStore

pub type RefStore = Store;

The standard type for a store to handle git references.

Aliased Type

struct RefStore {
    pub write_reflog: WriteReflog,
    pub namespace: Option<Namespace>,
    pub prohibit_windows_device_names: bool,
    pub precompose_unicode: bool,
    /* private fields */
}

Fields
write_reflog: WriteReflog

The way to handle reflog edits

namespace: Option<Namespace>

The namespace to use for edits and reads

prohibit_windows_device_names: bool

This is only needed on Windows, where some device names are reserved at any level of a path, so that reading or writing refs/heads/CON for example would read from the console, or write to it.

precompose_unicode: bool

If set, we will convert decomposed unicode like a\u308 into precomposed unicode like ä when reading ref names from disk. Note that this is an internal operation that isn’t observable on the outside, but it’s needed for lookups to packed-refs or symlinks to work correctly. Iterated references will be returned verbatim, thus when sending them over the wire they have to be precomposed as needed.