gix

Module worktree

Modules

attributes

excludes

glob

Provide glob Patterns for matching against paths or anything else.

ignore

Parse .gitignore files and provide utilities to match against them.

index

Feature Flags

object

This crate provides types for read-only git objects backed by bytes provided in git’s serialization format as well as mutable versions of these. Both types of objects can be encoded.

open_index

pathspec

proxy

stack

Structs

Proxy

A stand-in to a worktree as result of a worktree iteration.

Stack

A cache for efficiently executing operations on directories and files which are encountered in sorted order. That way, these operations can be re-used for subsequent invocations in the same directory.

Enums

IndexPersistedOrInMemory

A type to represent an index which either was loaded from disk as it was persisted there, or created on the fly in memory.

Type Aliases

Index

A lazily loaded and auto-updated worktree index.