gix::worktree

Enum IndexPersistedOrInMemory

pub enum IndexPersistedOrInMemory {
    Persisted(Index),
    InMemory(File),
}

A type to represent an index which either was loaded from disk as it was persisted there, or created on the fly in memory.

Variants

Persisted(Index)

The index as loaded from disk, and shared across clones of the owning Repository.

InMemory(File)

A temporary index as created from the HEAD^{tree}, with the file path set to the place where it would be stored naturally.

Note that unless saved explicitly, it will not persist.

Implementations

impl IndexPersistedOrInMemory

pub fn into_owned(self) -> File

Consume this instance and turn it into an owned index file.

Note that this will cause the persisted index to be cloned, which would happen whenever the repository has a worktree.