gix::worktree::glob::search

Function add_patterns_file

pub fn add_patterns_file<T>(
    patterns: &mut Vec<List<T>>,
    source: PathBuf,
    follow_symlinks: bool,
    root: Option<&Path>,
    buf: &mut Vec<u8>,
) -> Result<bool, Error>
where
    T: Pattern,

Add the given file at source if it exists, otherwise do nothing. If a root is provided, it’s not considered a global file anymore. Returns true if the file was added, or false if it didn’t exist.