gix::worktree

Crate ignore

Parse .gitignore files and provide utilities to match against them.

Feature Flags

Modules

parse

search

Structs

Search

A grouping of lists of patterns while possibly keeping associated to their base path in order to find matches.

Enums

Kind

The kind of ignored item.

Functions

parse

Parse git ignore patterns, line by line, from bytes.