gix::worktree::ignore

Module search

Structs

Ignore

An implementation of the Pattern trait for ignore patterns.

Match

Describes a matching pattern within a search for ignored paths.

Functions

pattern_idx_matching_relative_path

Like pattern_matching_relative_path(), but returns an index to the pattern that matched relative_path, instead of the match itself.

pattern_matching_relative_path

Return a match if a pattern matches relative_path, providing a pre-computed basename_pos which is the starting position of the basename of relative_path. is_dir is true if relative_path is a directory. case specifies whether cases should be folded during matching or not.