gix::worktree::ignore::search

Function pattern_matching_relative_path

pub fn pattern_matching_relative_path<'a>(
    list: &'a List<Ignore>,
    relative_path: &BStr,
    basename_pos: Option<usize>,
    is_dir: Option<bool>,
    case: Case,
) -> Option<Match<'a>>

Return a match if a pattern matches relative_path, providing a pre-computed basename_pos which is the starting position of the basename of relative_path. is_dir is true if relative_path is a directory. case specifies whether cases should be folded during matching or not.