gix::worktree

Crate index

Feature Flags

Modules

decode

entry

extension

file

fs

This module contains a Metadata implementation that must be used instead of std::fs::Metadata to assure that the ctime information is populated exactly like the one in git, which wouldn’t be the case on unix.

hash

This crate provides types for identifying git objects using a hash digest.

init

validate

Validation for various kinds of git related items.

verify

write

Structs

AccelerateLookup

A backing store for accelerating lookups of entries in a case-sensitive and case-insensitive manner.

Entry

An entry in the index, identifying a non-tree item on disk.

File

An index file whose state was read from a file on disk.

State

An in-memory cache of a fully parsed git index file.

Enums

Version

All known versions of a git index file.

Type Aliases

PathStorage

The type to use and store paths to all entries.

PathStorageRef

The type to use and store paths to all entries, as reference