gix::worktree::index::decode

Struct Options

pub struct Options {
    pub thread_limit: Option<usize>,
    pub min_extension_block_in_bytes_for_threading: usize,
    pub expected_checksum: Option<ObjectId>,
}

Options to define how to decode an index state from bytes.

Fields
thread_limit: Option<usize>

If Some(_), we are allowed to use more than one thread. If Some(N), use no more than N threads. If Some(0)|None, use as many threads as there are logical cores.

This applies to loading extensions in parallel to entries if the common EOIE extension is available. It also allows to use multiple threads for loading entries if the IEOT extension is present.

min_extension_block_in_bytes_for_threading: usize

The minimum size in bytes to load extensions in their own thread, assuming there is enough num_threads available. If set to 0, for example, extensions will always be read in their own thread if enough threads are available.

expected_checksum: Option<ObjectId>

Set the expected hash of this index if we are read as part of a link extension.

We will abort reading this file if it doesnโt match.