gix::worktree::index::extension

Module untracked_cache

Structs

Directory

A directory with information about its untracked files, and its sub-directories

OidStat

A structure to track filesystem stat information along with an object id, linking a worktree file with what’s in our ODB.

Constants

SIGNATURE

Only used as an indicator

Functions

decode

Decode an untracked cache extension from data, assuming object hashes are of type object_hash.