gix::worktree::index

Module fs

This module contains a Metadata implementation that must be used instead of std::fs::Metadata to assure that the ctime information is populated exactly like the one in git, which wouldn’t be the case on unix.

Structs

Metadata

A structure to partially mirror std::fs::Metadata.