gix::worktree::index::validate::path

Function component_is_windows_device

pub fn component_is_windows_device(input: &BStr) -> bool

Return true if the path component at input looks like a Windows device, like CON or LPT1 (case-insensitively).

This is relevant only on Windows, where one may be tricked into reading or writing to such devices. When reading from CON, a console-program may block until the user provided input.