gix::worktree::object

Enum ObjectRef

pub enum ObjectRef<'a> {
    Tree(TreeRef<'a>),
    Blob(BlobRef<'a>),
    Commit(CommitRef<'a>),
    Tag(TagRef<'a>),
}

Immutable objects are read-only structures referencing most data from a byte slice.

Immutable objects are expected to be deserialized from bytes that acts as backing store, and they cannot be mutated or serialized. Instead, one will convert them into their mutable counterparts which support mutation and serialization.

An ObjectRef is representing Trees, Blobs, Commits, or Tags.

Variants

Tree(TreeRef<'a>)

Blob(BlobRef<'a>)

Commit(CommitRef<'a>)

Tag(TagRef<'a>)

Implementations

impl<'a> ObjectRef<'a>

pub fn from_loose(data: &'a [u8]) -> Result<ObjectRef<'a>, LooseDecodeError>

Deserialize an object from a loose serialisation

pub fn from_bytes(kind: Kind, data: &'a [u8]) -> Result<ObjectRef<'a>, Error>

Deserialize an object of kind from the given data.

pub fn into_owned(self) -> Object

Convert the immutable object into a mutable version, consuming the source in the process.

Note that this is an expensive operation.

pub fn to_owned(&self) -> Object

Convert this immutable object into its mutable counterpart.

Note that this is an expensive operation.

impl<'a> ObjectRef<'a>

Convenient access to contained objects.

pub fn as_blob(&self) -> Option<&BlobRef<'a>>

Interpret this object as blob.

pub fn into_blob(self) -> Option<BlobRef<'a>>

Interpret this object as blob, chainable.

pub fn as_commit(&self) -> Option<&CommitRef<'a>>

Interpret this object as commit.

pub fn into_commit(self) -> Option<CommitRef<'a>>

Interpret this object as commit, chainable.

pub fn as_tree(&self) -> Option<&TreeRef<'a>>

Interpret this object as tree.

pub fn into_tree(self) -> Option<TreeRef<'a>>

Interpret this object as tree, chainable

pub fn as_tag(&self) -> Option<&TagRef<'a>>

Interpret this object as tag.

pub fn into_tag(self) -> Option<TagRef<'a>>

Interpret this object as tag, chainable.

pub fn kind(&self) -> Kind

Return the kind of object.