gix::worktree::object

Struct Commit

pub struct Commit {
    pub tree: ObjectId,
    pub parents: SmallVec<[ObjectId; 1]>,
    pub author: Signature,
    pub committer: Signature,
    pub encoding: Option<BString>,
    pub message: BString,
    pub extra_headers: Vec<(BString, BString)>,
}

A mutable git commit, representing an annotated state of a working tree along with a reference to its historical commits.

Fields
tree: ObjectId

The hash of recorded working tree state.

parents: SmallVec<[ObjectId; 1]>

Hash of each parent commit. Empty for the first commit in repository.

author: Signature

Who wrote this commit.

committer: Signature

Who committed this commit.

This may be different from the author in case the author couldn’t write to the repository themselves and is commonly encountered with contributed commits.

encoding: Option<BString>

The name of the message encoding, otherwise UTF-8 should be assumed.

message: BString

The commit message documenting the change.

extra_headers: Vec<(BString, BString)>

Extra header fields, in order of them being encountered, made accessible with the iterator returned by extra_headers().

Implementations

impl Commit

pub fn extra_headers( &self, ) -> ExtraHeaders<impl Iterator<Item = (&BStr, &BStr)>>

Returns a convenient iterator over all extra headers.