gix::worktree::object

Trait Exists

pub trait Exists {
    // Required method
    fn exists(&self, id: &oid) -> bool;
}

Check if an object is present in an object store.

Required Methods

fn exists(&self, id: &oid) -> bool

Returns true if the object exists in the database.