gix::worktree::object

Trait Write

pub trait Write {
    // Required method
    fn write_stream(
        &self,
        kind: Kind,
        size: u64,
        from: &mut dyn Read,
    ) -> Result<ObjectId, Box<dyn Error + Send + Sync>>;

    // Provided methods
    fn write(
        &self,
        object: &dyn WriteTo,
    ) -> Result<ObjectId, Box<dyn Error + Send + Sync>> { ... }
    fn write_buf(
        &self,
        object: Kind,
        from: &[u8],
    ) -> Result<ObjectId, Box<dyn Error + Send + Sync>> { ... }
}

Describe the capability to write git objects into an object store.

Required Methods

fn write_stream( &self, kind: Kind, size: u64, from: &mut dyn Read, ) -> Result<ObjectId, Box<dyn Error + Send + Sync>>

As write, but takes an input stream. This is commonly used for writing blobs directly without reading them to memory first.

Provided Methods

fn write( &self, object: &dyn WriteTo, ) -> Result<ObjectId, Box<dyn Error + Send + Sync>>

Write objects using the intrinsic kind of hash into the database, returning id to reference it in subsequent reads.

fn write_buf( &self, object: Kind, from: &[u8], ) -> Result<ObjectId, Box<dyn Error + Send + Sync>>

As write, but takes an object kind along with its encoded bytes.