gix::worktree::object

Trait WriteTo

pub trait WriteTo {
    // Required methods
    fn write_to(&self, out: &mut dyn Write) -> Result<(), Error>;
    fn kind(&self) -> Kind;
    fn size(&self) -> u64;

    // Provided method
    fn loose_header(&self) -> SmallVec<[u8; 28]>  { ... }
}

Writing of objects to a Write implementation

Required Methods

fn write_to(&self, out: &mut dyn Write) -> Result<(), Error>

Write a representation of this instance to out.

fn kind(&self) -> Kind

Returns the type of this object.

fn size(&self) -> u64

Returns the size of this object’s representation (the amount of data which would be written by write_to).

size’s value has no bearing on the validity of the object, as such it’s possible for size to return a sensible value but write_to to fail because the object was not actually valid in some way.

Provided Methods

fn loose_header(&self) -> SmallVec<[u8; 28]>

Returns a loose object header based on the object’s data