gix::worktree::object

Module tree

Modules

editor

write

Structs

Editor

The state needed to apply edits instantly to in-memory trees.

Entry

An entry in a Tree, similar to an entry in a directory.

EntryMode

The mode of items storable in a tree, similar to the file mode on a unix file system.

EntryRef

An element of a TreeRef.

Enums

EntryKind

A discretized version of ideal and valid values for entry modes.