gix::worktree::stack

Struct Platform

pub struct Platform<'a> { /* private fields */ }

Implementations

impl<'a> Platform<'a>

Access

pub fn path(&self) -> &'a Path

The full path to relative will be returned for use on the file system.

pub fn is_excluded(&self) -> bool

See if the currently set entry is excluded as per exclude and git-ignore files.

Note that this threats both classes, trashable and precious as equal. If you need to differentiate, use matching_exclude_pattern() or excluded_kind().

Panics

If the cache was configured without exclude patterns.

pub fn excluded_kind(&self) -> Option<Kind>

See if a non-negative ignore-pattern matches and obtain the kind of exclude, or return None if the path isn’t excluded.

This is similar to is_excluded(), but provides details that are useful to decide what to do with the excluded item.

pub fn matching_exclude_pattern(&self) -> Option<Match<'_>>

Check all exclude patterns to see if the currently set path matches any of them.

Note that this pattern might be negated, and means this path in included.

Panics

If the cache was configured without exclude patterns.

pub fn matching_attributes(&self, out: &mut Outcome) -> bool

Match all attributes at the current path and store the result in out, returning true if at least one attribute was found.

Panics

If the cache was configured without attributes.