gix

Struct Worktree

pub struct Worktree<'repo> { /* private fields */ }

A worktree checkout containing the files of the repository in consumable form.

Implementations

impl Worktree<'_>

pub fn open_index(&self) -> Result<File, Error>

A shortcut to crate::Repository::open_index().

pub fn index(&self) -> Result<Index, Error>

A shortcut to crate::Repository::index().

impl Worktree<'_>

pub fn excludes( &self, overrides: Option<Search>, ) -> Result<AttributeStack<'_>, Error>

Configure a file-system cache checking if files below the repository are excluded.

This takes into consideration all the usual repository configuration, namely:

$XDG_CONFIG_HOME/โฆ/ignore if core.excludesFile is not set, otherwise use the configured file.

$GIT_DIR/info/exclude if present.

When only excludes are desired, this is the most efficient way to obtain them. Otherwise use Worktree::attributes() for accessing both attributes and excludes.

impl<'repo> Worktree<'repo>

pub fn attributes( &self, overrides: Option<Search>, ) -> Result<AttributeStack<'repo>, Error>

Configure a file-system cache checking if files below the repository are excluded or for querying their attributes.

This takes into consideration all the usual repository configuration, namely:

$XDG_CONFIG_HOME/โฆ/ignore|attributes if core.excludesFile|attributesFile is not set, otherwise use the configured file.

$GIT_DIR/info/exclude|attributes if present.

pub fn attributes_only(&self) -> Result<AttributeStack<'repo>, Error>

Like attributes(), but without access to exclude/ignore information.

impl<'repo> Worktree<'repo>

pub fn pathspec( &self, patterns: impl IntoIterator<Item = impl AsRef<BStr>>, ) -> Result<Pathspec<'repo>, Error>

Configure pathspecs patterns to be matched against, with pathspec attributes read from the worktree and then from the index if needed.

Note that the empty_patterns_match_prefix flag of the parent method defaults to true.

Deviation

Pathspec attributes match case-insensitively by default if the underlying filesystem is configured that way.

impl<'repo> Worktree<'repo>

Access

pub fn base(&self) -> &'repo Path

Read the location of the checkout, the base of the work tree

pub fn is_main(&self) -> bool

Return true if this worktree is the main worktree associated with a non-bare git repository.

It cannot be removed.

pub fn is_locked(&self) -> bool

Return true if this worktree cannot be pruned, moved or deleted, which is useful if it is located on an external storage device.

Always false for the main worktree.

pub fn lock_reason(&self) -> Option<BString>

Provide a reason for the locking of this worktree, if it is locked at all.

Note that we squelch errors in case the file cannot be read in which case the reason is an empty string.

pub fn id(&self) -> Option<&BStr>

Return the ID of the repository worktree, if it is a linked worktree, or None if itโs a linked worktree.

pub fn dot_git_exists(&self) -> bool

Returns true if the .git file or directory exists within the worktree.

This is an indicator for the worktree to be checked out particularly if the parent repository is a submodule.